% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{as.data.frame.LongTable}
\alias{as.data.frame.LongTable}
\title{Coerce a LongTable to a data.frame}
\usage{
\method{as.data.frame}{LongTable}(x)
}
\arguments{
\item{x}{[`LongTable`] to coerce to `data.frame`.}
}
\value{
[`data.frame`] containing the data from the LongTable, with the
  `LongTable.config' attribute containg the metadata needed to reverse
   the coercion operation.
}
\description{
Coerce a LongTable to a data.frame
}
