% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coerce.R
\name{as.long.table}
\alias{as.long.table}
\title{Coerce to data.table to LongTable}
\usage{
as.long.table(x)
}
\arguments{
\item{x}{A [`data.table`] with the 'LongTable.config' attribute, containing
three lists named assayCols, rowDataCols and colDataCols. This attribute is
automatically created when coercing from a LongTable to a data.table.}
}
\value{
[`LongTable`] object configured with the LongTable.config
}
\description{
Coerce a data.table with the proper configuration attributes back to a LongTable
}
