% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assay.R
\name{assay,LongTable,ANY-method}
\alias{assay,LongTable,ANY-method}
\title{Retrieve an assay `data.table` object from the `assays` slot of a `LongTable`
   object}
\usage{
\S4method{assay}{LongTable,ANY}(x, i, withDimnames = FALSE, metadata = FALSE, key = TRUE)
}
\arguments{
\item{x}{[`LongTable`] The `LongTable` object to get the assay from.}

\item{i}{[`integer`] or [`character`] vector containing the index or name
of the assay, respectively.}

\item{withDimnames}{[`logical`] Should the dimension names be returned
joined to the assay. This retrieves both the row and column identifiers
and returns them attached to the}

\item{metadata}{[`logical`] Should all of the metadata also be joined to
the assay. This is useful when modifying assays as the resulting list
has all the information needed to recreated the LongTable object.}

\item{key}{[`logical`] Should the key columns also be returned?}
}
\description{
Retrieve an assay `data.table` object from the `assays` slot of a `LongTable`
   object
}
