% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assay.R
\name{assay<-,LongTable,character-method}
\alias{assay<-,LongTable,character-method}
\title{Add or replace an assay in a LongTable by name}
\usage{
\S4method{assay}{LongTable,character}(x, i) <- value
}
\arguments{
\item{x}{[`LongTable`]}

\item{i}{[`character`]}

\item{value}{[`data.frame`] or [`data.table`]}
}
\value{
[`LongTable`] With updated assays slot.
}
\description{
Add or replace an assay in a LongTable by name
}
