% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{assayCols,LongTable-method}
\alias{assayCols,LongTable-method}
\title{Retrieve the value columns for the assays in a LongTable}
\usage{
\S4method{assayCols}{LongTable}(object, i)
}
\arguments{
\item{object}{[`LongTable`]}

\item{i}{Optional parameter specifying the [`character`] name or [`interger`]
index of the assay to get the column names for. If missing, returns a
list of value column names for all the assays.}
}
\value{
A [`list`] of `character` vectors containing the value column names for
    each assay if i is missing, otherwise a `character` vector of value column
    names for the selected assay.
}
\description{
Retrieve the value columns for the assays in a LongTable
}
