% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assayNames.R
\name{assayNames,LongTable-method}
\alias{assayNames,LongTable-method}
\title{Retrieve the assay names from a `LongTable` object.}
\usage{
\S4method{assayNames}{LongTable}(x)
}
\arguments{
\item{x}{A [`LongTable`] object to retrieve the assay names from.}
}
\value{
[`character`] Names of the assays contained in the `LongTable`.
}
\description{
Retrieve the assay names from a `LongTable` object.
}
