% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assays.R
\name{assays,LongTable-method}
\alias{assays,LongTable-method}
\title{Return a list of `data.table` objects with the assay measurements from a
 `LongTable` object.}
\usage{
\S4method{assays}{LongTable}(x, withDimnames = FALSE, metadata = FALSE, key = TRUE)
}
\arguments{
\item{x}{[`LongTable`] What to extract the assay data from.}

\item{withDimnames}{[`logical`] Should the returned assays be joined to
the row and column identifiers (i.e., the pseudo dimnames of the object).}

\item{metadata}{[`logical`] Should row and column metadata also be joined
to the returned assays. This is useful for modifying assays before
reconstructing a new LongTable.}
}
\value{

}
\description{
Return a list of `data.table` objects with the assay measurements from a
 `LongTable` object.
}
