% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-assays.R
\name{assays<-,LongTable,list-method}
\alias{assays<-,LongTable,list-method}
\title{Setter method for the assays slot in a LongTable object}
\usage{
\S4method{assays}{LongTable,list}(x) <- value
}
\arguments{
\item{x}{A [`LongTable`] to modify the assays in.}

\item{value}{A [`list`] of `data.frame` or `data.table` objects, all of which
contain the row and column identifiers and metadata.}
}
\value{
A copy of the [`LongTable`] with the assays modified
}
\description{
Setter method for the assays slot in a LongTable object
}
