% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-buildLongTable.R
\name{buildLongTable,data.frame-method}
\alias{buildLongTable,data.frame-method}
\title{Create a LongTable object from a single data.table or data.frame object}
\usage{
\S4method{buildLongTable}{data.frame}(from, rowDataCols, colDataCols, assayCols)
}
\arguments{
\item{from}{[`character`] Path to the .csv file containing the data and
metadata from which to build the `LongTable`.}

\item{rowDataCols}{[`list`] List with two `character` vectors, the first
specifying one or more columns to be used as cell identifiers (e.g.,
cell-line name columns) and the second containing any additional metadata
columns related to the cell identifiers. If you wish to rename any of
these columns, assign the new names to their respective character vectors.}

\item{colDataCols}{[`list`] List with two `character` vectors, the first
specifying one or more columns to be used as column identifiers (e.g.,
drug name columns) and the second containing any additional metadata
columns related to the column identifiers. If you wish to rename any of
these columns, assign the new names to their respective character vectors.}

\item{assayCols}{[`list`] A named list of character vectors specifying how to
parse assay columns into a list of `data.table`s. Each list data.table
will be named for the name of corresponding list item and contain the columns
specified in the character vector of column names in each list item. If
there are no names for assayCols, the assays will be numbered by instead.}
}
\value{
A [`LongTable`] object containing one or more assays, indexed by
  rowID and colID.
}
\description{
Create a LongTable object from a single data.table or data.frame object
}
