% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-buildLongTable.R
\name{buildLongTable,list-method}
\alias{buildLongTable,list-method}
\title{Create a LongTable object from a list containing file paths, data.frames and
  data.tables.}
\usage{
\S4method{buildLongTable}{list}(from, rowDataCols, colDataCols, assayCols)
}
\arguments{
\item{from}{[`list`] A list containing any combination of character file paths,
data.tables and data.frames which will be used to construct the LongTable.}
}
\description{
Create a LongTable object from a list containing file paths, data.frames and
  data.tables.
}
