% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{buildLongTableFromCSV}
\alias{buildLongTableFromCSV}
\title{Create a LongTable object from a single .csv file}
\usage{
buildLongTableFromCSV(filePath, rowDataCols, colDataCols, assayCols)
}
\arguments{
\item{filePath}{[`character`] Path to the .csv file containing the data and
metadata from which to build the `LongTable`.}

\item{rowDataCols}{[`list`] List with two `character` vectors, the first
specifying one or more columns to be used as cell identifiers (e.g.,
cell-line name columns) and the second containing any additional metadata
columns related to the cell identifiers.}

\item{colDataCols}{[`list`] List with two `character` vectors, the first
specifying one or more columns to be used as column identifiers (e.g.,
drug name columns) and the second containing any additional metadata
columns related to the column identifiers.}

\item{assayCols}{[`list`] A named list of character vectors specifying how to
parse assay columns into a list of `data.table`s. Each list data.table
will be named for the name of corresponding list item and contain the columns
specified in the character vector of column names in each list item.}
}
\value{
A [`LongTable`] object containing one or more assays, indexed by
  rowID and colID.
}
\description{
Create a LongTable object from a single .csv file
}
