% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCLEsmall}
\alias{CCLEsmall}
\title{Cancer Cell Line Encyclopedia (CCLE) Example PharmacoSet}
\format{
PharmacoSet object
}
\usage{
data(CCLEsmall)
}
\description{
A small example version of the CCLE PharmacoSet, used in the
documentation examples. All credit for the data goes to the CCLE
group at the Broad Institute. This is not a full version of the dataset; most of
the dataset was removed to make runnable example code. For the full dataset,
please download using the downloadPSet function.
}
\references{
Barretina et al. The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity. Nature, 2012
}
\keyword{datasets}
