% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cellInfo.R
\name{cellInfo<-,PharmacoSet,data.frame-method}
\alias{cellInfo<-,PharmacoSet,data.frame-method}
\title{cellInfo<- Generic}
\usage{
\S4method{cellInfo}{PharmacoSet,data.frame}(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace cell info in}

\item{value}{A \code{data.frame} with the new cell annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for cellInfo replace method
}
\examples{
data(CCLEsmall)
cellInfo(CCLEsmall) <- cellInfo(CCLEsmall)

}
