% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cellNames<-}
\alias{cellNames<-}
\title{cellNames<- Generic}
\usage{
cellNames(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to update}

\item{value}{A \code{character} vector of the new cell names}
}
\value{
Updated \code{PharmacoSet}
}
\description{
A generic for the cellNames replacement method
}
\examples{
data(CCLEsmall)
cellNames(CCLEsmall) <- cellNames(CCLEsmall)

}
