% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CGPsmall}
\alias{CGPsmall}
\title{Cancer Genome Project Example PharmacoSet}
\format{PharmacoSet object}
\usage{
data(CGPsmall)
}
\description{
A small example version of the Cancer Genome Project PharmacoSet, used in the
documentation examples. All credit for the data goes to the Cancer Genome Project
group at the Sanger.
}
\references{
Garnett et al. Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature, 2012.
}
\keyword{dataset}

