% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-class.R
\name{checkPsetStructure}
\alias{checkPsetStructure}
\title{A function to verify the structure of a PharmacoSet}
\usage{
checkPsetStructure(object, plotDist = FALSE, result.dir = ".")
}
\arguments{
\item{object}{A \code{PharmacoSet} to be verified}

\item{plotDist}{Should the function also plot the distribution of molecular data?}

\item{result.dir}{The path to the directory for saving the plots as a string}
}
\value{
Prints out messages whenever describing the errors found in the
structure of the object object passed in.
}
\description{
This function checks the structure of a PharamcoSet, ensuring that the
correct annotations are in place and all the required slots are filled so
that matching of cells and drugs can be properly done across different types
of data and with other studies.
}
\examples{
data(CCLEsmall)
checkPsetStructure(CCLEsmall)

}
