% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cnvData<-}
\alias{cnvData<-}
\title{cnvData<- Generic}
\usage{
cnvData(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace cnv data in}

\item{value}{A \code{matrix} with the new cnv data}
}
\description{
Generic for cnvData replace method
}

