% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cnvData}
\alias{cnvData}
\title{cnvData Generic}
\usage{
cnvData(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve cnv data from}
}
\value{
a \code{data.frame} with the cnv experiment data
}
\description{
Generic for cnvData method
}

