% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cnvInfo<-}
\alias{cnvInfo<-}
\title{cnvInfo<- Generic}
\usage{
cnvInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace cnv annotations in}

\item{value}{A \code{data.frame} with the new cnv annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for cnvInfo replace method
}

