% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cnvInfo}
\alias{cnvInfo}
\title{cnvInfo Generic}
\usage{
cnvInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve cnv annotations from}
}
\value{
a \code{data.frame} with the cnv annotations
}
\description{
Generic for cnvInfo method
}

