% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\name{colData,LongTable-method}
\alias{colData,LongTable-method}
\title{Retrieve the column metadata table from a LongTable object}
\usage{
\S4method{colData}{LongTable}(x, key = FALSE)
}
\arguments{
\item{x}{A [`LongTable`] to retrieve column metadata from.}

\item{key}{[`logical`] She the colKey column also be returned? Defaults to
FALSE.}
}
\value{
A [`data.table`] containing row identifiers and metadata.
}
\description{
Retrieve the column metadata table from a LongTable object
}
