% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-colData.R
\name{colData<-,LongTable,ANY-method}
\alias{colData<-,LongTable,ANY-method}
\title{Updates the `colData` slot as long as the ID columns are not changed.}
\usage{
\S4method{colData}{LongTable,ANY}(x) <- value
}
\arguments{
\item{x}{A [`LongTable`] object to modify.}

\item{value}{A [`data.table`] or [`data.frame`] to update with.}
}
\value{
A copy of the [`LongTable`] object with the `colData`
  slot updated.
}
\description{
Updates the `colData` slot as long as the ID columns are not changed.
}
