% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{colIDs,LongTable-method}
\alias{colIDs,LongTable-method}
\title{Get the id column names for the colData slot of a LongTable}
\usage{
\S4method{colIDs}{LongTable}(object, data = FALSE, key = FALSE)
}
\arguments{
\item{object}{A [`LongTable`] to get the colData id columns for.}

\item{data}{[`logical`] Should the colData for the id columns be returned
instead of the column names? Default is FALSE.}

\item{key}{[`logical`] Should the key column also be returned? Default is FALSE.}
}
\value{
A [`character`] vector of colData column names if data is FALSE,
     otherwise a [`data.table`] with the data from the colData id columns.
}
\description{
Get the id column names for the colData slot of a LongTable
}
