% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dimnames.R
\name{colnames,LongTable-method}
\alias{colnames,LongTable-method}
\title{Get the column names from a `LongTable` object.}
\usage{
\S4method{colnames}{LongTable}(x)
}
\arguments{
\item{x}{A [`LongTable`] object to get the column names from}
}
\value{
[`character`] Vector of column names.
}
\description{
Get the column names from a `LongTable` object.
}
