% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{computeBliss}
\alias{computeBliss}
\title{Compute Bliss Null References}
\usage{
computeBliss(viability_1, viability_2)
}
\arguments{
\item{viability_1}{\code{numeric} monotherapeutic response of treatment 1.}

\item{viability_2}{\code{numeric} monotherapeutic response of treatment 2.}
}
\value{
\code{numeric} expected response of the two treatments combined
under Bliss null assumption.
}
\description{
Given two \code{numeric} containing viability of two monotherapy respectively,
Compute Bliss null reference values for the expected response
of the two treatments combined.
}
\examples{
(bliss <- computeBliss(0.75, 0.65))

}
