% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{computeDeltaScore}
\alias{computeDeltaScore}
\title{Compute ZIP delta score}
\usage{
computeDeltaScore(
  viability,
  treatment1dose,
  treatment2dose,
  HS_1,
  HS_2,
  EC50_1,
  EC50_2
)
}
\arguments{
\item{viability}{\code{numeric} Observed viability of two treatments.}

\item{treatment1dose}{\code{numeric} a vector of concentrations for treatment 1}

\item{treatment2dose}{\code{numeric} a vector of concentrations for treatment 2}

\item{HS_1}{\code{numeric} Hill coefficient of treatment 1}

\item{HS_2}{\code{numeric} Hill coefficient of treatment 2}

\item{EC50_1}{\code{numeric} relative EC50 of treatment 1.}

\item{EC50_2}{\code{numeric} relative EC50 of treatment 2.}
}
\value{
\code{numeric} ZIP delta score to quantify synergy.
}
\description{
Compute ZIP delta score as described in the original paper.
}
