% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{computeZIP}
\alias{computeZIP}
\title{Computes ZIP Null References}
\usage{
computeZIP(treatment1dose, treatment2dose, HS_1, HS_2, EC50_1, EC50_2)
}
\arguments{
\item{treatment1dose}{\code{numeric} a vector of concentrations for treatment 1}

\item{treatment2dose}{\code{numeric} a vector of concentrations for treatment 2}

\item{HS_1}{\code{numeric} Hill coefficient of treatment 1}

\item{HS_2}{\code{numeric} Hill coefficient of treatment 2}

\item{EC50_1}{\code{numeric} relative EC50 of treatment 1.}

\item{EC50_2}{\code{numeric} relative EC50 of treatment 2.}
}
\value{
\code{numeric} expected viability under ZIP null assumption.
}
\description{
Predict the additive response of a treatment combination under
the ZIP null assumption.
}
