% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{computeZIPdelta}
\alias{computeZIPdelta}
\title{Generic to compute ZIP delta scores from an S4 object}
\usage{
computeZIPdelta(object, ...)
}
\arguments{
\item{object}{\code{S4} An object to compute delta scores from.}

\item{...}{Allow new arguments to this generic.}
}
\value{
Depends on the implemented method.
}
\description{
Generic to compute ZIP delta scores from an S4 object
}
\examples{
print("Generics shouldn't need examples?")

}
