% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPsetMolecularProfilesToSE.R
\name{convertPsetMolecularProfilesToSE}
\alias{convertPsetMolecularProfilesToSE}
\title{PSet molecularProfiles from ESets to SEs}
\usage{
convertPsetMolecularProfilesToSE(pSet)
}
\arguments{
\item{pSet}{\code{S4} A PharmacoSet containing molecular data in ExpressionSets}
}
\value{
\code{S4} A PharmacoSet containing molecular data in a SummarizedExperiments
}
\description{
Converts all ExpressionSet objects within the molecularProfiles slot of a 
  PharmacoSet to SummarizedExperiments
}
