% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dim.R
\name{dim,LongTable-method}
\alias{dim,LongTable-method}
\title{Get the dimensions of a `LongTable` object.}
\usage{
\S4method{dim}{LongTable}(x)
}
\arguments{
\item{x}{A [`LongTable`] object to retrieve dimensions for.}
}
\value{
[`numeric`] Vector of object dimensions.
}
\description{
Get the dimensions of a `LongTable` object.
}
