% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{methods}
\name{dim,PharmacoSet-method}
\alias{dim,PharmacoSet-method}
\title{Get the dimensions of a PharmacoSet}
\usage{
\S4method{dim}{PharmacoSet}(x)
}
\arguments{
\item{x}{PharmacoSet}
}
\value{
A named vector with the number of Cells and Drugs in the PharmacoSet
}
\description{
Get the dimensions of a PharmacoSet
}
