% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{dimnames,LongTable-method}
\alias{dimnames,LongTable-method}
\title{Getter for the dimnames of a `LongTable` object}
\usage{
\S4method{dimnames}{LongTable}(x)
}
\arguments{
\item{x}{The [`LongTable`] object to retrieve the dimnames for}
}
\value{
[`list`] List with two character vectors, one for row and one for
    column names.
}
\description{
Getter for the dimnames of a `LongTable` object
}
