% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{dnaData<-}
\alias{dnaData<-}
\title{dnaData<- Generic}
\usage{
dnaData(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace dna data in}

\item{value}{A \code{matrix} with the new dna data}
}
\description{
Generic for dnaData replace method
}

