% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{dnaData}
\alias{dnaData}
\title{dnaData Generic}
\usage{
dnaData(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve dna data from}
}
\value{
a \code{data.frame} with the dna experiment data
}
\description{
Generic for dnaData method
}

