% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{dnaInfo<-}
\alias{dnaInfo<-}
\title{dnaInfo<- Generic}
\usage{
dnaInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace dna info in}

\item{value}{A \code{data.frame} with the new dna annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for dnaInfo replace method
}

