% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{dnaInfo}
\alias{dnaInfo}
\title{dnaInfo Generic}
\usage{
dnaInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve dna annotations from}
}
\value{
a \code{data.frame} with the dna annotations
}
\description{
Generic for dnaInfo method
}

