% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.colIDData,LongTable-method}
\alias{.colIDData,LongTable-method}
\title{Extract the column ID columns from colData of a LongTable}
\usage{
\S4method{.colIDData}{LongTable}(object, key = TRUE)
}
\arguments{
\item{object}{[`LongTable`] What to get the colIDData from}

\item{key}{[`logical`] Should the colKey also be returned? Default is TRUE.}
}
\description{
Extract the column ID columns from colData of a LongTable
}
\keyword{internal}
