% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.colIDData}
\alias{.colIDData}
\title{Return the identifiers for the column meta data in an object}
\usage{
.colIDData(object, ...)
}
\description{
Return the identifiers for the column meta data in an object
}
