% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.colIDs,LongTable-method}
\alias{.colIDs,LongTable-method}
\title{Developer accessor method to determine which columns hold the column identifiers
  in the `colData` slot of a `LongTable` object.}
\usage{
\S4method{.colIDs}{LongTable}(object)
}
\arguments{
\item{A}{[`object`] `LongTable` object to retrieve the column identifier
column indexes from.}
}
\value{
[`numeric`] A numeric vector of integer column indexes for the ID
  columns in `colData`.
}
\description{
Developer accessor method to determine which columns hold the column identifiers
  in the `colData` slot of a `LongTable` object.
}
\keyword{internal}
