% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.colIDs}
\alias{.colIDs}
\title{Private method to retrieve the .colIDs property from an object}
\usage{
.colIDs(object, ...)
}
\description{
Private method to retrieve the .colIDs property from an object
}
