% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.dimIDs,LongTable-method}
\alias{.dimIDs,LongTable-method}
\title{Developer accessor method to determine which columns hold the row identifiers
  in the `rowData` slot of a `LongTable` object.}
\usage{
\S4method{.dimIDs}{LongTable}(object)
}
\arguments{
\item{A}{[`LongTable`] object to retrieve the row identifier column
indexes from.}
}
\value{
A [`list`] containing two `numeric` vectors, the first for the
   identifier columns for the `rowData` slot and the second with the same
   for the `colData` slot
}
\description{
Developer accessor method to determine which columns hold the row identifiers
  in the `rowData` slot of a `LongTable` object.
}
\keyword{internal}
