% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.dimIDs}
\alias{.dimIDs}
\title{Private method to retrieve the both the .rowIDs and .colIDs properties from
  an object in a list.}
\usage{
.dimIDs(object, ...)
}
\arguments{
\item{object}{[`any`] An object with a .intern slot with the items .rowIDs
and .colIDs}
}
\value{
A [`list`] v
}
\description{
Private method to retrieve the both the .rowIDs and .colIDs properties from
  an object in a list.
}
\keyword{internal}
