% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{.filterLongDataTable}
\alias{.filterLongDataTable}
\title{Filter a data.table object based on the rowID and colID columns}
\usage{
.filterLongDataTable(DT, indexList)
}
\arguments{
\item{DT}{[`data.table`] Object with the columns rowID and colID, preferably
as the key columns.}

\item{indexList}{[`list`] Two integer vectors, one indicating the rowIDs and
one indicating the colIDs to filter the `data.table` on.}
}
\value{
[`data.table`] A copy of `DT` subset on the row and column IDs specified
 in `indexList`.
}
\description{
Filter a data.table object based on the rowID and colID columns
}
\keyword{internal}
