% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\name{.row_IDs_are_identical}
\alias{.row_IDs_are_identical}
\title{Helper to determine if the rowData rowIDs are identical between a new
  rowData data.table object and an existing `LongTable`}
\usage{
.row_IDs_are_identical(rowIDs, newRowData)
}
\description{
Helper to determine if the rowData rowIDs are identical between a new
  rowData data.table object and an existing `LongTable`
}
\keyword{internal}
