% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.rowIDData,LongTable-method}
\alias{.rowIDData,LongTable-method}
\title{Extract the row ID columns from `rowDat` of a `LongTable`}
\usage{
\S4method{.rowIDData}{LongTable}(object, key = TRUE)
}
\arguments{
\item{object}{[`LongTable`] What to get the colIDData from.}

\item{key}{[`logical`] Should the `colKey` column also be returned? Default
is TRUE.}
}
\description{
Extract the row ID columns from `rowDat` of a `LongTable`
}
\keyword{internal}
