% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.rowIDData}
\alias{.rowIDData}
\title{Return the identifiers for the row metadata columns in an object}
\usage{
.rowIDData(object, ...)
}
\description{
Return the identifiers for the row metadata columns in an object
}
