% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.rowIDs,LongTable-method}
\alias{.rowIDs,LongTable-method}
\title{Developer accessor method to determine which columns hold the row identifiers
  in the `rowData` slot of a `LongTable` object.}
\usage{
\S4method{.rowIDs}{LongTable}(object)
}
\arguments{
\item{A}{[`object`] `LongTable` object to retrieve the row identifier column
indexes from.}
}
\value{
[`numeric`] A numeric vector of integer column indexes for the ID
  columns in `rowData`.
}
\description{
Developer accessor method to determine which columns hold the row identifiers
  in the `rowData` slot of a `LongTable` object.
}
\keyword{internal}
