% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.rowIDs}
\alias{.rowIDs}
\title{Private method to retrieve the .rowIDs property from an object}
\usage{
.rowIDs(object, ...)
}
\description{
Private method to retrieve the .rowIDs property from an object
}
\keyword{internal}
