% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{.selectDataTable}
\alias{.selectDataTable}
\title{Select a set of column names from a data.table, returning a copy of the
  data.table with duplicate rows removed}
\usage{
.selectDataTable(colNames, DT, keep.rownames = FALSE)
}
\arguments{
\item{colNames}{[`character`] The column names to select from the data.table}

\item{DT}{[`data.table`, `data.frame`, `matrix`] An object coercible to a `data.table`.
Please note rownames will be dropped by default.}

\item{keep.rownames}{[`logical` or `character`] Passed through to the data.table coercing if DT is not a
`data.table`. If TRUE, rownames will be caputured in the `rn` column; if FALSE (default) rownames will
be dropped; if `character`, rownames will be captured in a column with the same name.}
}
\value{
[`data.table`] Copy of `DT` containing only the specified columns, with duplicate rows removed.
}
\description{
Select a set of column names from a data.table, returning a copy of the
  data.table with duplicate rows removed
}
\keyword{internal}
