% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.sensitivitySlotToLongTable}
\alias{.sensitivitySlotToLongTable}
\title{Reconstruct the}
\usage{
.sensitivitySlotToLongTable(object)
}
\arguments{
\item{object}{[`LongTable`]}
}
\description{
Reconstruct the
}
