% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{.variableToCodeString}
\alias{.variableToCodeString}
\title{Convert an R object in a variable into a string of the code necessary to
  create that object}
\usage{
.variableToCodeString(variable)
}
\arguments{
\item{variable}{[`Symbol`] A symbol containing an R variable}
}
\value{
[`string`] A string representation of the code necessary to
  reconstruct the variable.
}
\description{
Convert an R object in a variable into a string of the code necessary to
  create that object
}
\keyword{internal}
