% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadSignatures.R
\name{downloadSignatures}
\alias{downloadSignatures}
\title{Download Drug Signatures}
\usage{
downloadSignatures(name = c("CGP", "CCLE", "CMAP"),
  download.method = "auto", saveDir = "./PSets/Sigs/", myfn = NULL)
}
\arguments{
\item{name}{\code{Character} string, the name of the PhamracoSet for which to
download signatures. The available options are CGP, CCLE, and CMAP}

\item{download.method}{\code{Character} string, the method used by
\code{download.file}. Defaults to \code{auto}.}

\item{saveDir}{\code{Character} string with the folder path where the
PharmacoSet should be saved. Defaults to \code{'./PSets/Sigs/'}. Will create
directory if it does not exist.}

\item{myfn}{\code{character} string, the file name to save the dataset under}
}
\description{
This function allows you to download an array of drug signatures, as would be
computed by the \code{drugPerturbationSig} and \code{drugSensitivitySig}
functions, for available perturbation and sensitivity
\code{PharmacoSets} respectively. This function allows the user to skip these very
lengthy calculation steps for the datasets available, and start their
analysis from the already computed signatures
}

