% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{drugInfo<- Generic}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{\link{PharmacoSet}} to replace drug info}

\item{value}{A \code{\link{data.frame}} with the new drug annotations}
}
\value{
The \code{\link{object}} with updated drug annotations
}
\description{
Generic for drugInfo replace method
}
