% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{drugNames}
\alias{drugNames}
\title{drugNames Generic}
\usage{
drugNames(object)
}
\arguments{
\item{object}{The \code{\link{PharmacoSet}} to return drug names from}
}
\value{
A \code{\link{character}} vector of drug names in the object
}
\description{
A generic for the drugNames method
}
\examples{
data(CCLEsmall)
drugNames(CCLEsmall)

}
