% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{estimateProjParams}
\alias{estimateProjParams}
\title{Estimate projected potency and shape parameter}
\usage{
estimateProjParams(
  dose_to,
  viability,
  dose_add,
  EC50_add,
  HS_add,
  E_inf_add = 0,
  use_L2 = FALSE,
  show_Rsqr = FALSE
)
}
\arguments{
\item{dose_to}{\code{numeric} a vector of concentrations of the drug being added to}

\item{viability}{\code{numeric} Observed viability of two treatments; target for fitting curve.}

\item{dose_add}{\code{numeric} a vector of concentrations of the drug added.}

\item{EC50_add}{\code{numeric} relative EC50 of the drug added.}

\item{HS_add}{\code{numeric} Hill coefficient of the drug added.}

\item{E_inf_add}{\code{numeric} Efficacy of the drug added.}

\item{use_L2}{\code{logical} Whether to use L2 loss for fitting curves.
This method produces cruder estimates for projected Hill parameters of
drug combinations difficult to optimise, but also faster to compute results.
Use it only if the default method is not progressing. Default \code{FALSE}.}

\item{show_Rsqr}{\code{logical} whether to show goodness-of-fit value in the result.}
}
\value{
\code{list}
* \code{EC50_proj}: Projected potency after adding a drug
* \code{HS_proj}: Projected Hill coefficient after adding a drug
}
\description{
Estimate the projected potency EC50 and the shape parameter HS
in the dose-response curve of a drug after adding another drug to it
by fitting a 2-parameter dose-response curve.
It assumes \eqn{E_min = 1} for the drug being added and
\eqn{E_inf = 0} for both drugs.
}
\examples{
(EC50_and_HS <- estimateProjParams(
  viability=c(0.9, 0.8, 0.7),
  dose_to=c(0.1, 0.01, 0.001),
  dose_add=c(1, 0.1, 0.01),
  EC50_add=0.1,
  HS_add=1.1,
  E_inf_add=0.1
))

}
