% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{featureInfo}
\alias{featureInfo}
\title{featureInfo Generic}
\usage{
featureInfo(pSet, mDataType)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve feature annotations from}

\item{mDataType}{the type of molecular data}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for featureInfo method
}
\examples{
data(CCLEsmall)
featureInfo(CCLEsmall, "rna")

}
