% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{featureNames}
\alias{featureNames}
\title{featureNames Generic}
\usage{
featureNames(pSet, mDataType)
}
\arguments{
\item{pSet}{The \code{PharmacoSet}}

\item{mDataType}{The molecular data type to return feature names for}
}
\value{
A \code{character} vector of the feature names
}
\description{
A generic for the featureNames method
}

