% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{fNames<-,PharmacoSet,character,character-method}
\alias{fNames<-,PharmacoSet,character,character-method}
\title{fNames<-}
\usage{
\S4method{fNames}{PharmacoSet,character,character}(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} object to update}

\item{mDataType}{The molecular data type to update}

\item{value}{A \code{character} vector of the new cell names}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Setter for the feature names of a \code{SummarizedExperiment} in the 
  molecularProfiles slot
}
\examples{
data(CCLEsmall)
fNames(CCLEsmall, 'rna') <- fNames(CCLEsmall, 'rna')

Set feature names of the object object

}
