% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{fNames<-,PharmacoSet,character-method}
\alias{fNames<-,PharmacoSet,character-method}
\title{fNames<-}
\usage{
\S4method{fNames}{PharmacoSet,character}(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} object to update}

\item{value}{A \code{character} vector of the new cell names}
}
\value{
Updated \code{PhmaSet}
}
\description{
Set feature names of the object object
}
