% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDrugSensitivity.R
\name{geneDrugSensitivity}
\alias{geneDrugSensitivity}
\title{Calcualte The Gene Drug Sensitivity}
\usage{
geneDrugSensitivity(
  x,
  type,
  batch,
  drugpheno,
  interaction.typexgene = FALSE,
  model = FALSE,
  standardize = c("SD", "rescale", "none"),
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{numeric} vector of gene expression values}

\item{type}{A \code{vector} of factors specifying the cell lines or type types}

\item{batch}{A \code{vector} of factors specifying the batch}

\item{drugpheno}{A \code{numeric} vector of drug sensitivity values (e.g.,
IC50 or AUC)}

\item{interaction.typexgene}{\code{boolean} Should interaction between gene
expression and cell/type type be computed? Default set to FALSE}

\item{model}{\code{boolean} Should the full linear model be returned? Default
set to FALSE}

\item{standardize}{\code{character} One of 'SD', 'rescale' or 'none'}

\item{verbose}{\code{boolean} Should the function display messages?}
}
\value{
A \code{vector} reporting the effect size (estimate of the coefficient
of drug concentration), standard error (se), sample size (n), t statistic,
and F statistics and its corresponding p-value.
}
\description{
TODO:: Write a description!
}
\examples{
print("TODO::")

}
