% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{geneInfo<-}
\alias{geneInfo<-}
\title{geneInfo<- Generic}
\usage{
geneInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace gene annotations in}

\item{value}{A \code{data.frame} with the new gene annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for geneInfo replace method
}

