% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{geneInfo}
\alias{geneInfo}
\title{geneInfo Generic}
\usage{
geneInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve gene annotations from}
}
\value{
a \code{data.frame} with the gene annotations
}
\description{
Generic for geneInfo method
}

