% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{geneNames}
\alias{geneNames}
\title{geneNames Generic}
\usage{
geneNames(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to return gene names from}
}
\value{
A vector of the gene names used in the PharmacoSet
}
\description{
A generic for the geneNames method
}

