% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HDAC_genes}
\alias{HDAC_genes}
\title{HDAC Gene Signature}
\format{a 13x2 data.frame with gene identifiers in the first column and
  direction change in the second}
\usage{
data(HDAC_genes)
}
\description{
A gene signature for HDAC inhibitors, as detailed by Glaser et al. The 
signature is mapped from the probe to gene level using
\code{probeGeneMapping}
}
\references{
Glaser et al. Gene expression profiling of multiple histone deacetylase (HDAC) inhibitors: defining a common gene set produced by HDAC inhibition in T24 and MDA carcinoma cell lines. Molecular cancer therapeutics, 2003.
}
\keyword{datasets}
