% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{Hill_4par}
\alias{Hill_4par}
\title{Predict viability from a 4-paramter Hill curve}
\usage{
Hill_4par(dose, HS, E_nnf, E_inf, EC50)
}
\arguments{
\item{dose}{\code{numeric} a vector of concentrations of the treatment in \code{log10} scale.}

\item{HS}{\code{numeric} Hill coefficient characterised by cooperativity and molecularity of
ligand-receptor binding of the treatment.}

\item{E_nnf}{\code{numeric} cellular viability under
the effect of the treatment at the minimum dose level.
In predicting monotherapeutic viability,
a sensible \code{E_nnf} should be the negative control value of 1.}

\item{E_inf}{\code{numeric} efficacy of the treatment;
viability produced by the maximum effect of the treatment.}

\item{EC50}{\code{numeric} potency of the treatment in \code{log10} scale;
the concentration necessary to cause half of the treatment efficacy.}
}
\value{
\code{numeric} predicted cellular viability values of the treatment
given \code{dose} administered respectively.
}
\description{
Viability predicted by 4-parameter logistic regression.
}
