% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectList.R
\name{intersectList}
\alias{intersectList}
\title{Utility to find the intersection between a list of more than two vectors or
lists}
\usage{
intersectList(...)
}
\arguments{
\item{...}{A list of or any number of vector like objects of the same mode,
which could also be operated on by the native R set operations}
}
\value{
A vector like object of the same mode as the first argument,
  containing only the intersection common to all arguments to the function
}
\description{
This function extends the native intersect function to work on two or more
arguments.
}
\examples{
list1 <- list('a', 'b', 'c')
list2 <- list('a', 'c')
list3 <- list('a', 'c', 'd')
listAll <- intersectList(list1, list2, list3)
listAll

}
