% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R, R/methods-[[.R
\docType{class}
\name{LongTable-class}
\alias{LongTable-class}
\alias{.LongTable}
\alias{[[,LongTable,ANY,ANY-method}
\title{LongTable class definition}
\usage{
\S4method{[[}{LongTable,ANY,ANY}(x, i, keys = FALSE)
}
\arguments{
\item{x}{[`LongTable`] object to retrieve assays from}

\item{i}{[`character`] name or [`integer`] index of the desired assay.}

\item{keys}{[`logical`] Should the row and column keys also be returned?
Defaults to FALSE to make applying aggregate functions more convenient.}

\item{drugs}{[`data.table`]}

\item{cells}{[`data.table`]}

\item{assays}{[`list`]}

\item{metadata}{[`list`]}
}
\value{
[`LongTable`] object containing the assay data from a
}
\description{
Define a private constructor method to be used to build a `LongTable` object.

Select an assay from within a LongTable object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: 
}}

\keyword{internal}
