% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{LongTable}
\alias{LongTable}
\title{LongTable constructor method}
\usage{
LongTable(
  rowData,
  rowIDs,
  colData,
  colIDs,
  assays,
  metadata = list(),
  keep.rownames = FALSE
)
}
\arguments{
\item{rowData}{[`data.table`, `data.frame`, `matrix`] A table like object
coercible to a `data.table` containing the a unique `rowID` column which
is used to key assays, as well as additional row metadata to subset on.}

\item{rowIDs}{[`character`, `integer`] A vector specifying
the names or integer indexes of the row data identifier columns. These
columns will be pasted together to make up the row.names of the
`LongTable` object.}

\item{colData}{[`data.table`, `data.frame`, `matrix`] A table like object
coercible to a `data.table` containing the a unique `colID` column which
is used to key assays, as well as additional column metadata to subset on.}

\item{colIDs}{[`character`, `integer`] A vector specifying
the names or integer indexes of the col data identifier columns. These
columns will be pasted together to make up the col.names of the
`LongTable` object.}

\item{assays}{A [`list`] containing one or more objects coercible to a
`data.table`, and keyed by rowID and colID corresponding to the rowID and
colID columns in colData and rowData.}

\item{metadata}{A [`list`] of metadata associated with the `LongTable`
object being constructed}

\item{keep.rownames}{[`logical` or `character`] Logical: whether rownames
should be added as a column if coercing to a `data.table`, default is FALSE.
If TRUE, rownames are added to the column 'rn'. Character: specify a custom
column name to store the rownames in.}
}
\value{
[`LongTable`] object
}
\description{
Constrcuts a long table #FIXME:: Better description
}
