% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{mDataNames,PharmacoSet-method}
\alias{mDataNames,PharmacoSet-method}
\title{mDataNames}
\usage{
\S4method{mDataNames}{PharmacoSet}(object)
}
\arguments{
\item{object}{PharamcoSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the PharmacoSet.
}
\examples{
data(CCLEsmall)
mDataNames(CCLEsmall)

}
