% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{mDataNames}
\alias{mDataNames}
\title{mDataNames}
\usage{
mDataNames(pSet)
}
\arguments{
\item{pSet}{PharamcoSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the PharmacoSet.
}
\examples{
data(CCLEsmall)
mDataNames(CCLEsmall)

}
