% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{metadata,LongTable-method}
\alias{metadata,LongTable-method}
\title{Getter method for the metadata slot of a `LongTable` object}
\usage{
\S4method{metadata}{LongTable}(x)
}
\arguments{
\item{x}{The [`LongTable`] object from which to retrieve the metadata list.}
}
\value{
[`list`] The contents of the `metadata` slot of the `LongTable`
  object.
}
\description{
Getter method for the metadata slot of a `LongTable` object
}
