% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-metadata.R
\name{metadata<-,LongTable-method}
\alias{metadata<-,LongTable-method}
\title{Setter method for the metadata slot of a `LongTable` object}
\usage{
\S4method{metadata}{LongTable}(x) <- value
}
\arguments{
\item{x}{[`LongTable`] The LongTable to update}

\item{value}{[`list`] A list of new metadata associated with a `LongTable`
object.}
}
\value{
[`LongTable`] A copy of the `LongTable` object with the `value` in
  the metadata slot.
}
\description{
Setter method for the metadata slot of a `LongTable` object
}
