% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{pertNumber}
\alias{pertNumber}
\title{pertNumber Generic}
\usage{
pertNumber(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet}}
}
\value{
A 3D \code{array} with the number of perturbation experiments per drug and cell line, and data type
}
\description{
A generic for the pertNumber method
}
\examples{
data(CCLEsmall)
pertNumber(CCLEsmall)

}
