% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-utils.R
\name{PharmacoSet-utils}
\alias{PharmacoSet-utils}
\title{Utility methods for a \code{PharmacoSet} object.}
\arguments{
\item{x}{A \code{PharmacoSet} object.}

\item{samples}{\code{character()} vector of sample names. Must be valid rownames
from \code{cellInfo(x)}.}

\item{features}{\code{character()} vector of feature names. Must be valid feature
names for a given \code{mDataType}}
}
\value{
See details.
}
\description{
Documentation for utility methods for a \code{PharmacoSet} object, such as
set operations like subset and intersect. See @details for information
on different types of methods and their implementations.
}
\examples{
data(CCLEsmall)
}
