% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{phenoInfo}
\alias{phenoInfo}
\title{phenoInfo Generic}
\usage{
phenoInfo(pSet, mDataType)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve rna annotations from}

\item{mDataType}{the type of molecular data}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for phenoInfo method
}
\examples{
data(CCLEsmall)
phenoInfo(CCLEsmall, mDataType="rna")

}
