% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{projectedResponse}
\alias{projectedResponse}
\title{Compute response of adding one treatment to the other}
\usage{
projectedResponse(dose_add, EC50_add, HS_add, dose_to, EC50_proj, HS_proj)
}
\arguments{
\item{dose_add}{\code{numeric} a vector of concentrations of the drug added.}

\item{EC50_add}{\code{numeric} relative EC50 of the drug added.}

\item{HS_add}{\code{numeric} Hill coefficient of the drug added.}

\item{dose_to}{\code{numeric} a vector of concentrations of the drug being added to}

\item{EC50_proj}{\code{numeric} changed relative EC50 of the drug being added to; the projected potency.}

\item{HS_proj}{\code{numeric} changed Hill coefficient of the drug being added to; the projected shape parameter.}
}
\value{
\code{numeric} viability after adding one drug to the other.
}
\description{
Response of adding one drug to the other,
with the assumption that E_inf = 0, and
E_min is dictated by the response of the drug added.
}
