% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{projViability}
\alias{projViability}
\title{Compute the projected viability after adding one treatment to the other}
\usage{
projViability(dose_to, EC50_proj, HS_proj, E_min_proj)
}
\arguments{
\item{dose_to}{\code{numeric} a vector of concentrations of the drug being added to}

\item{EC50_proj}{\code{numeric} changed relative EC50 of the drug being added to; the projected potency.}

\item{HS_proj}{\code{numeric} changed Hill coefficient of the drug being added to; the projected shape parameter.}

\item{E_min_proj}{\code{numeric} Projected \code{E_min} given by
the viability of the added treatment at a fixed dose.}
}
\value{
\code{numeric} viability after adding one drug to the other.
}
\description{
Viability of adding one drug to the other,
with the assumption that E_inf_add = 0, and
E_min of the drug being added to is dictated by the response of the drug added.
}
