% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{pSetName}
\alias{pSetName}
\title{pSetName Generic}
\usage{
pSetName(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet}}
}
\value{
The name of the PharmacoSet
}
\description{
A generic for the pSetName method
}

