% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{reindex}
\alias{reindex}
\title{Generic method for resetting indexing in an S4 object}
\usage{
reindex(object, ...)
}
\arguments{
\item{object}{[`S4`] An object to redo indexing for}

\item{...}{[`pairlist`] Allow definition of new parameters to this generic.}
}
\description{
This method allows integer indexes used to maintain referential integrity
  internal to an S4 object to be reset. This is useful particularly after
  subsetting an object, as certain indexes may no longer be present in the
  object data. Reindexing removes gaps integer indexes and ensures that the
  smallest contiguous integer values are used in an objects indexes.
}
