% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{rnaData<-}
\alias{rnaData<-}
\title{rnaData<- Generic}
\usage{
rnaData(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace rna data in}

\item{value}{A \code{matrix} with the new rna data}
}
\description{
Generic for rnaData replace method
}

