% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{rnaData}
\alias{rnaData}
\title{rnaData Generic}
\usage{
rnaData(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve rna data from}
}
\value{
a \code{data.frame} with the rna experiment data
}
\description{
Generic for rnaData method
}

