% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{rnaInfo<-}
\alias{rnaInfo<-}
\title{rnaInfo<- Generic}
\usage{
rnaInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace rna annotations in}

\item{value}{A \code{data.frame} with the new rna annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for rnaInfo replace method
}

