% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{rnaInfo}
\alias{rnaInfo}
\title{rnaInfo Generic}
\usage{
rnaInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve rna annotations from}
}
\value{
a \code{data.frame} with the rna annotations
}
\description{
Generic for rnaInfo method
}

