% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\name{rowData,LongTable-method}
\alias{rowData,LongTable-method}
\title{Retrieve the row metadata table from a LongTable object}
\usage{
\S4method{rowData}{LongTable}(x, key = FALSE, use.names = FALSE)
}
\arguments{
\item{x}{A [`LongTable`] object to retrieve the row metadata from.}

\item{key}{[`logical`] Should the rowKey column also be returned? Defaults
to FALSE.}

\item{use.names}{[`logical`] This parameter is just here to stop matching
they key argument to use.names from the rowData generic. It doesn't do
anything.}
}
\value{
A [`data.table`] containing rowID, row identifiers, and row metadata.
}
\description{
Retrieve the row metadata table from a LongTable object
}
