% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-rowData.R
\name{rowData<-,LongTable-method}
\alias{rowData<-,LongTable-method}
\title{Updates the `rowData` slot as long as the ID columns are not changed.}
\usage{
\S4method{rowData}{LongTable}(x) <- value
}
\arguments{
\item{x}{A [`LongTable`] object to modify.}

\item{value}{A [`data.table`] or [`data.frame`] to update}
}
\value{
A copy of the [`LongTable`] object with the `rowData`
  slot updated.
}
\description{
Updates the `rowData` slot as long as the ID columns are not changed.
}
