% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{rowIDs,LongTable-method}
\alias{rowIDs,LongTable-method}
\title{Get the id column names for the rowData slot of a LongTable}
\usage{
\S4method{rowIDs}{LongTable}(object, data = FALSE, key = FALSE)
}
\arguments{
\item{object}{A [`LongTable`] to get the rowData id columns for.}

\item{data}{[`logical`] Should the rowData for the id columns be returned
instead of the column names? Default is FALSE.}

\item{key}{[`logical`] Should the key column also be returned?}
}
\value{
A [`character`] vector of rowData column names if data is FALSE,
     otherwise a [`data.table`] with the data from the rowData id columns.
}
\description{
Get the id column names for the rowData slot of a LongTable
}
