% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{rowMeta,LongTable-method}
\alias{rowMeta,LongTable-method}
\title{Get the id column names for the rowData slot of a LongTable}
\usage{
\S4method{rowMeta}{LongTable}(object, data = FALSE, key = FALSE)
}
\arguments{
\item{object}{A [`LongTable`] to get the rowData metadata columns for.}

\item{data}{[`logical`] Should the rowData for the metadata columns be returned
instead of the column names? Default is FALSE.}

\item{key}{[`logical`] Should the key column also be returned? Default is FALSE}
}
\value{
A [`character`] vector of rowData column names if data is FALSE,
     otherwise a [`data.table`] with the data from the rowData metadta columns.
}
\description{
Get the id column names for the rowData slot of a LongTable
}
