% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{rownames,LongTable-method}
\alias{rownames,LongTable-method}
\title{Get the row names from a `LongTable` object.}
\usage{
\S4method{rownames}{LongTable}(x)
}
\arguments{
\item{x}{A [`LongTable`] object to get the row names from}
}
\value{
[`character`] Vector of row names.
}
\description{
Get the row names from a `LongTable` object.
}
