% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityData<-}
\alias{sensitivityData<-}
\title{sensitivityData<- Generic}
\usage{
sensitivityData(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace sensitivity data in}

\item{value}{A \code{matrix} with the new sensitivity data}
}
\description{
Generic for sensitivityData replace method
}

