% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityData}
\alias{sensitivityData}
\title{sensitivityData Generic}
\usage{
sensitivityData(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve sensitivity data from}
}
\value{
a \code{data.frame} with the sensitivity experiment data
}
\description{
Generic for sensitivityData method
}

