% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityInfo<-}
\alias{sensitivityInfo<-}
\title{sensitivityInfo<- Generic}
\usage{
sensitivityInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace sensitivity experiment annotations in}

\item{value}{A \code{data.frame} with the new sensitivity annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for sensitivityInfo replace method
}

