% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityInfo}
\alias{sensitivityInfo}
\title{sensitivityInfo Generic}
\usage{
sensitivityInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve sensitivity experiment annotations from}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityInfo method
}
\examples{
data(CCLEsmall)
sensitivityInfo(CCLEsmall)

}
