% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityMeasures}
\alias{sensitivityMeasures}
\title{sensitivityMeasures Generic}
\usage{
sensitivityMeasures(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet}}
}
\value{
A \code{character} vector of available sensitivity measures
}
\description{
A generic for the sensitivity measures function
}

