% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityProfiles}
\alias{sensitivityProfiles}
\title{sensitivityProfiles Generic}
\usage{
sensitivityProfiles(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve sensitivity experiment data from}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for sensitivityProfiles method
}
\examples{
data(CCLEsmall)
sensitivityProfiles(CCLEsmall)

}
