% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sensitivitySlotToLongTable.R
\name{sensitivitySlotToLongTable,PharmacoSet-method}
\alias{sensitivitySlotToLongTable,PharmacoSet-method}
\title{Reconstruct the data in the @sensitivity slot list into a LongTable object.}
\usage{
\S4method{sensitivitySlotToLongTable}{PharmacoSet}(object)
}
\arguments{
\item{object}{A [`PharmacoSet`] with a list in the sensitivity slot containing
items raw, profiles, info and n.}
}
\value{
A `LongTable` with the data from the sensitivity slot.
}
\description{
Reconstruct the data in the @sensitivity slot list into a LongTable object.
}
