% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensNumber<-}
\alias{sensNumber<-}
\title{sensNumber<- Generic}
\usage{
sensNumber(object) <- value
}
\arguments{
\item{object}{A \code{PharmacoSet}}

\item{value}{A new \code{data.frame} with the number of sensitivity experiments per drug and cell line}
}
\value{
The updated \code{PharmacoSet}
}
\description{
A generic for the sensNumber method
}
\examples{
data(CCLEsmall)
sensNumber(CCLEsmall) <- sensNumber(CCLEsmall)

}
