% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensNumber}
\alias{sensNumber}
\title{sensNumber Generic}
\usage{
sensNumber(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet}}
}
\value{
A \code{data.frame} with the number of sensitivity experiments per drug and cell line
}
\description{
A generic for the sensNumber method
}
\examples{
data(CCLEsmall)
sensNumber(CCLEsmall)

}
