% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-LongTable.R
\name{show,LongTable-method}
\alias{show,LongTable-method}
\title{Show method for the LongTable class}
\usage{
\S4method{show}{LongTable}(object)
}
\arguments{
\item{object}{A [`LongTable`] object to print the results for.}
}
\description{
Show method for the LongTable class
}
