% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signatureClass.R
\docType{methods}
\name{show,PharmacoGxSignatures-method}
\alias{show,PharmacoGxSignatures-method}
\title{Show PharmacoGx Signatures}
\usage{
\S4method{show}{PharmacoGxSignatures}(object)
}
\arguments{
\item{object}{\code{PharmacoGxSignatures}}
}
\description{
Show PharmacoGx Signatures
}

