% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\docType{methods}
\name{show,PharmacoSig-method}
\alias{show,PharmacoSig-method}
\title{Show PharmacoGx Signatures}
\usage{
\S4method{show}{PharmacoSig}(object)
}
\arguments{
\item{object}{\code{PharmacoSig}}
}
\description{
Show PharmacoGx Signatures
}

