% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signatureClass.R
\name{showSigAnnot}
\alias{showSigAnnot}
\title{Show the Annotations of a signature object}
\usage{
showSigAnnot(Sigs)
}
\arguments{
\item{Sigs}{An object of the \code{PharmacoSig} Class, as
returned by \code{drugPerturbationSig} or \code{drugSensitivitySig}}
}
\description{
Show the Annotations of a signature object
}

