% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{snpData<-}
\alias{snpData<-}
\title{snpData<- Generic}
\usage{
snpData(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace snp data in}

\item{value}{A \code{matrix} with the new snp data}
}
\description{
Generic for snpData replace method
}

