% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{snpData}
\alias{snpData}
\title{snpData Generic}
\usage{
snpData(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve snp data from}
}
\value{
a \code{data.frame} with the snp experiment data
}
\description{
Generic for snpData method
}

