% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{snpInfo<-}
\alias{snpInfo<-}
\title{snpInfo<- Generic}
\usage{
snpInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace snp annotations in}

\item{value}{A \code{data.frame} with the new snp annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for snpInfo replace method
}

