% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{snpInfo}
\alias{snpInfo}
\title{snpInfo Generic}
\usage{
snpInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve snp annotations from}
}
\value{
a \code{data.frame} with the snp annotations
}
\description{
Generic for snpInfo method
}

