% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-[.R
\name{[,LongTable,ANY,ANY,ANY-method}
\alias{[,LongTable,ANY,ANY,ANY-method}
\title{`[`}
\usage{
\S4method{[}{LongTable,ANY,ANY,ANY}(x, i, j, assays, ..., drop = FALSE)
}
\arguments{
\item{x}{object}

\item{i}{row query}

\item{j}{col query}

\item{assays}{[`character`]}

\item{...}{ad args}

\item{drop}{does nothing}
}
\description{
`[`
}
