% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-[.R
\name{[,PharmacoSet,ANY,ANY,ANY-method}
\alias{[,PharmacoSet,ANY,ANY,ANY-method}
\title{\code{[}}
\usage{
\S4method{[}{PharmacoSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{object}

\item{i}{Cell lines to keep in object}

\item{j}{Drugs to keep in object}

\item{...}{further arguments}

\item{drop}{A boolean flag of whether to drop single dimensions or not}
}
\value{
Returns the subsetted object
}
\description{
\code{[}
}
\examples{
data(CCLEsmall)
CCLEsmall["WM1799", "Sorafenib"]

}
