% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{methods}
\name{[,PharmacoSet-method}
\alias{[,PharmacoSet-method}
\title{`[` 
 @param x numeric
 @param i numeric
 @param j numeric
 @param ... further arguments
 @param drop A boolean flag of whether to drop single dimensions or not
 @export}
\usage{
\S4method{[}{PharmacoSet}(x, i, j, ..., drop = FALSE)
}
\description{
`[` 
 @param x numeric
 @param i numeric
 @param j numeric
 @param ... further arguments
 @param drop A boolean flag of whether to drop single dimensions or not
 @export
}

