% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionList.R
\name{unionList}
\alias{unionList}
\title{Utility to find the union between a list of more than two vectors or
lists}
\usage{
unionList(...)
}
\arguments{
\item{...}{A list of or any number of vector like objects of the same mode,
which could also be operated on by the native R set operations}
}
\value{
A vector like object of the same mode as the first argument,
  containing all the elements of all arguments passed to the function
}
\description{
This function extends the native union function to work on two or more
arguments.
}
\examples{
list1 <- list('a', 'b')
list2 <- list('a', 'c')
list3 <- list('c', 'd')
listAll <- unionList(list1, list2, list3)
listAll

}
