% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject-methods.R
\name{updateObject,PharmacoSet-method}
\alias{updateObject,PharmacoSet-method}
\title{Update the PharmacoSet class after changes in it struture or API}
\usage{
\S4method{updateObject}{PharmacoSet}(object)
}
\arguments{
\item{object}{A \code{PharmacoSet} object to update the class structure for.}
}
\value{
\code{PharmacoSet} with update class structure.
}
\description{
Update the PharmacoSet class after changes in it struture or API
}
