% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validatePSetMolecularProfilesToSEConversion.R
\name{validatePsetMolecularProfilesToSEConversion}
\alias{validatePsetMolecularProfilesToSEConversion}
\title{Validate PSet molecularProfiles Conversion}
\usage{
validatePsetMolecularProfilesToSEConversion(pSet_old, pSet_new)
}
\arguments{
\item{pSet}{[S4] a PSet containing molecularProfiles as ExpressionSets}
}
\value{
[message] Any slots which are not the same
}
\description{
Checks that all the information contained in an ExpressionSet molecularProfile 
  was successfully tranferred to the SummarizedExperiment molecularProfile
}
