% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPSet.R
\name{availablePSets}
\alias{availablePSets}
\title{Return a table of PharmacoSets available for download}
\usage{
availablePSets(saveDir = file.path(".", "PSets"), myfn = "PSets.csv",
  verbose = TRUE)
}
\arguments{
\item{saveDir}{\code{character} Directory to save the table of PSets}

\item{myfn}{\code{character} The filename for the table of PSets}

\item{verbose}{\code{bool} Should status messages be printed during download.}
}
\value{
A data.frame with details about the available PharmacoSet objects
}
\description{
The function fetches a table of all PharmacoSets available for download from the PharmacoGx server. 
The table includes the names of the PharamcoSet, the types of data available
in the object, and the date of last update.
}
\examples{
if (interactive()){
availablePSets()
}

}
