% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCLEsmall}
\alias{CCLEsmall}
\title{Cancer Cell Line Encyclopedia (CCLE) Example PharmacoSet}
\format{PharmacoSet object}
\usage{
data(CCLEsmall)
}
\description{
A small example version of the CCLE PharmacoSet, used in the
documentation examples. All credit for the data goes to the CCLE
group at the Broad Institute.
}
\references{
Barretina et al. The Cancer Cell Line Encyclopedia enables predictive modelling of anticancer drug sensitivity. Nature, 2012
}
\keyword{dataset}

