% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cellInfo}
\alias{cellInfo}
\title{cellInfo Generic}
\usage{
cellInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve cell info from}
}
\value{
a \code{data.frame} with the cell annotations
}
\description{
Generic for cellInfo method
}
\examples{
data(CCLEsmall)
cellInfo(CCLEsmall)

}
