% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{cellNames}
\alias{cellNames}
\title{cellNames Generic}
\usage{
cellNames(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to return cell names from}
}
\value{
A vector of the cell names used in the PharmacoSet
}
\description{
A generic for the cellNames method
}
\examples{
data(CCLEsmall)
cellNames(CCLEsmall)

}
