% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CMAPsmall}
\alias{CMAPsmall}
\title{Connectivity Map Example PharmacoSet}
\format{PharmacoSet object}
\usage{
data(CMAPsmall)
}
\description{
A small example version of the Connectivity Map PharmacoSet, used in the
documentation examples. All credit for the data goes to the Connectivity Map
group at the Broad Institute. This is not a full version of the dataset, most of
of the dataset was removed to make runnable example code. For the full dataset,
please download using the downloadPSet function.
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
