% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeABC.R
\name{computeABC}
\alias{computeABC}
\title{Fits dose-response curves to data given by the user
and returns the ABC of the fitted curves.}
\usage{
computeABC(conc1, conc2, viability1, viability2, trunc = TRUE)
}
\arguments{
\item{conc1}{[vector] is a vector of drug concentrations.}

\item{conc2}{[vector] is a vector of drug concentrations.}

\item{viability1}{[vector] is a vector whose entries are the viability values observed in the presence of the
drug concentrations whose logarithms are in the corresponding entries of conc1, expressed as percentages
of viability in the absence of any drug.}

\item{viability2}{[vector] is a vector whose entries are the viability values observed in the presence of the
drug concentrations whose logarithms are in the corresponding entries of conc2, expressed as percentages
of viability in the absence of any drug.}

\item{trunc}{[logical], if true, causes viability data to be truncated to lie between 0 and 1 before
curve-fitting is performed.}
}
\description{
Fits dose-response curves to data given by the user
and returns the ABC of the fitted curves.
}

