% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeAUC.R
\name{computeAUC}
\alias{computeAUC}
\title{Return AUC (Area Under the drug response curve) for an experiment of a pSet by taking
its concentration and viability as input.}
\usage{
computeAUC(concentration, viability, trunc = TRUE)
}
\arguments{
\item{concentration}{[vector] A concentration range that the AUC should be computed for that range.
Concentration range by default considered as not logarithmic scaled.}

\item{viability}{[vector] Viablities correspondant to the concentration range passed as first parameter.
The range of viablity values by definition should be between 0 and 100. But the viabalities greater than
100 and lower than 0 are also accepted.}

\item{trunc}{[binary] A flag that identify if the viabality values should be truncated to be in the
range of (0,100)}
}
\description{
Return AUC (Area Under the drug response curve) for an experiment of a pSet by taking
its concentration and viability as input.
}

