% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{computeHSA}
\alias{computeHSA}
\title{Compute HSA Null References}
\usage{
computeHSA(viability_1, viability_2)
}
\arguments{
\item{viability_1}{\code{numeric} monotherapeutic response of treatment 1.}

\item{viability_2}{\code{numeric} monotherapeutic response of treatment 2.}
}
\value{
\code{numeric} expected response of the two treatments combined
using the highest response of the two (lower viability).
}
\description{
Given two \code{numeric} containing viability of two monotherapy respectively,
Compute highest single-agent (HSA) values as the expected response
of the two treatments combined.
}
\examples{
(hsa <- computeHSA(0.75, 0.65))

}
