% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\docType{methods}
\name{computeZIPdelta,TreatmentResponseExperiment-method}
\alias{computeZIPdelta,TreatmentResponseExperiment-method}
\title{Compute ZIP delta score}
\usage{
\S4method{computeZIPdelta}{TreatmentResponseExperiment}(object, use_L2 = FALSE, nthread = 1L)
}
\arguments{
\item{object}{\link{TreatmentResponseExperiment}
The \code{TreatmentResponseExperiment} from which to extract assays
\code{mono_profile} and \code{combo_viability} to compute ZIP delta scores.}

\item{use_L2}{\code{logical}
whether to use L2-loss rather than the default optimisation method.
This method produces cruder estimates for projected Hill parameters of
drug combinations difficult to optimise, but also faster to compute results.
Use it only if the default method is not progressing. Default \code{FALSE}.}

\item{nthread}{\code{integer} Number of cores used to perform computation.
Default 1.}
}
\value{
\link{TreatmentResponseExperiment} with assay \code{combo_scores} containing \code{delta_scores}
}
\description{
Following the calculation of ZIP delta score as in Appendix A3.
See reference for details.

Compute ZIP delta score as described in the original paper.
}
\references{
Yadav, B., Wennerberg, K., Aittokallio, T., & Tang, J. (2015). Searching for Drug Synergy in Complex Dose–Response Landscapes Using an Interaction Potency Model. Computational and Structural Biotechnology Journal, 13, 504–513. https://doi.org/10.1016/j.csbj.2015.09.001
}
