% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{dateCreated}
\alias{dateCreated}
\title{dateCreated Generic}
\usage{
dateCreated(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet}}
}
\value{
The date the PharmacoSet was created
}
\description{
A generic for the dateCreated method
}
\examples{
data(CCLEsmall)
dateCreated(CCLEsmall)

}
