% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{.computeZIPdelta}
\alias{.computeZIPdelta}
\title{Vector-based version of \link{computeZIPdelta}}
\usage{
.computeZIPdelta(
  treatment1id,
  treatment2id,
  treatment1dose,
  treatment2dose,
  HS_1,
  HS_2,
  EC50_1,
  EC50_2,
  E_inf_1,
  E_inf_2,
  viability,
  residual = c("logcosh", "Cauchy", "normal"),
  nthread = 1L
)
}
\arguments{
\item{treatment1id}{\code{character} a vector of identifiers for treatment 1}

\item{treatment2id}{\code{character} a vector of identifiers for treatment 2}

\item{treatment1dose}{\code{numeric} a vector of concentrations for treatment 1}

\item{treatment2dose}{\code{numeric} a vector of concentrations for treatment 2}

\item{HS_1}{\code{numeric} Hill coefficient of treatment 1}

\item{HS_2}{\code{numeric} Hill coefficient of treatment 2}

\item{EC50_1}{\code{numeric} relative EC50 of treatment 1.}

\item{EC50_2}{\code{numeric} relative EC50 of treatment 2.}

\item{E_inf_1}{\code{numeric} viability produced by the maximum attainable effect of treatment 1.}

\item{E_inf_2}{\code{numeric} viability produced by the maximum attainable effect of treatment 2.}

\item{viability}{\code{numeric} Observed viability of the two treatments combined.}

\item{residual}{\code{character} Method used to minimise residual in fitting curves.
3 methods available: \code{logcosh}, \code{normal}, \code{Cauchy}.
The default method is \code{logcosh}.
It minimises the logarithmic hyperbolic cosine loss of the residuals
and provides the fastest estimation among the three methods,
with fitting quality in between \code{normal} and \code{Cauchy}.
The other two methods minimise residuals by
considering the truncated probability distribution (as in their names) for the residual.
\code{Cauchy} provides the best fitting quality but also takes the longest to run.}

\item{nthread}{\code{integer} Number of cores used to perform computation.
Default 1.}
}
\value{
\code{numeric} delta scores of every dose combinations for any given treatment combinations.
}
\description{
Following the calculation of ZIP delta score as in Appendix A3.
See reference for details.
}
\examples{
\dontrun{
combo_profiles <- CoreGx::buildComboProfiles(tre, c("HS", "EC50", "E_inf", "viability"))
combo_twowayFit <- fitTwowayZIP(combo_profiles)
combo_twowayFit |>
    aggregate(
        delta_score = .computeZIPdelta(
            treatment1id = treatment1id,
            treatment2id = treatment2id,
            treatment1dose = treatment1dose,
            treatment2dose = treatment2dose,
            HS_1 = HS_1, HS_2 = HS_2,
            EC50_1 = EC50_1, EC50_2 = EC50_2,
            E_inf_1 = E_inf_1, E_inf_2 = E_inf_2,
            viability = viability
        ),
        by = c("treatment1id", "treatment2id", "sampleid")
    )
}

}
\references{
Yadav, B., Wennerberg, K., Aittokallio, T., & Tang, J. (2015). Searching for Drug Synergy in Complex Dose–Response Landscapes Using an Interaction Potency Model. Computational and Structural Biotechnology Journal, 13, 504–513. https://doi.org/10.1016/j.csbj.2015.09.001
}
