% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.resaveAllExampleDatasets}
\alias{.resaveAllExampleDatasets}
\title{Utility function to resave all datasets after modifying convertPSetMolecularProfiles}
\usage{
.resaveAllExampleDatasets(datasets)
}
\arguments{
\item{datasets}{\code{character} A list of the example datasets to update}
}
\description{
Utility function to resave all datasets after modifying convertPSetMolecularProfiles
}
