% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-summarizeSensitivityProfiles.R
\name{.summarizeSensProfiles}
\alias{.summarizeSensProfiles}
\title{Summarize the sensitivity profiles when the sensitivity slot is a LongTable}
\usage{
.summarizeSensProfiles(
  object,
  sensitivity.measure = "auc_recomputed",
  cell.lines,
  drugs,
  summary.stat,
  fill.missing = TRUE
)
}
\description{
Summarize the sensitivity profiles when the sensitivity slot is a LongTable
}
\keyword{internal}
