% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{drugInfo<- Generic}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new drug annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for drugInfo replace method
}
\examples{
data(CCLEsmall)
drugInfo(CCLEsmall) <- drugInfo(CCLEsmall)

}
