% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Generic}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{The \code{PharmacoSet} to retrieve drug info from}
}
\value{
A [`data.frame`] of annotations for drugs in the object
}
\description{
Generic for drugInfo getter method
}
