% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Generic}
\usage{
drugInfo(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
Generic for drugInfo method
}
\examples{
data(CCLEsmall)
drugInfo(CCLEsmall)

}
