% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{drugNames}
\alias{drugNames}
\title{drugNames Generic}
\usage{
drugNames(object)
}
\arguments{
\item{object}{The [`PharmacoSet`] to return drug names from}
}
\value{
A [`character`] vector of drug names in the object
}
\description{
A generic for the drugNames method
}
\examples{
data(CCLEsmall)
drugNames(CCLEsmall)

}
