% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{drugNames}
\alias{drugNames}
\title{drugNames Generic}
\usage{
drugNames(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to return drug names from}
}
\value{
A vector of the drug names used in the PharmacoSet
}
\description{
A generic for the drugNames method
}
\examples{
data(CCLEsmall)
drugNames(CCLEsmall)

}
