% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{estimateNewPotency}
\alias{estimateNewPotency}
\title{Estimate projected potency and shape parameter}
\usage{
estimateNewPotency(viability, dose_add, EC50_add, HS_add, dose_to)
}
\arguments{
\item{viability}{\code{numeric} Observed viability of two treatments; target for fitting curve.}

\item{dose_add}{\code{numeric} a vector of concentrations of the drug added.}

\item{EC50_add}{\code{numeric} relative EC50 of the drug added.}

\item{HS_add}{\code{numeric} Hill coefficient of the drug added.}

\item{dose_to}{\code{numeric} a vector of concentrations of the drug being added to}
}
\value{
\code{list} \itemize{
\item{"EC50"}{Projected potency after adding a drug}
\item{"HS"}{Projected Hill coefficient after adding a drug}
}
}
\description{
Estimate the projected potency EC50 and the shape parameter HS
in the dose-response curve of a drug after adding another drug to it
by fitting a 2-parameter dose-response curve.
It assumes \eqn{E_min = 1} for the drug being added and
\eqn{E_inf = 0} for both drugs.
}
