% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSynergy.R
\name{fitTwowayZIP}
\alias{fitTwowayZIP}
\title{Two-way fitting for projected dose-response curve.}
\usage{
fitTwowayZIP(combo_profiles, use_L2 = FALSE, show_Rsqr = FALSE, nthread = 1L)
}
\arguments{
\item{combo_profiles}{\link{data.table} contains three parameters of dose-response curves
for each single agent in a drug comnbination,
and the observed viability of two treatments combined.}

\item{use_L2}{\code{logical}
whether to use L2-loss rather than the default optimisation method.
This method produces cruder estimates for projected Hill parameters of
drug combinations difficult to optimise, but also faster to compute results.
Use it only if the default method is not progressing. Default \code{FALSE}.}

\item{show_Rsqr}{\code{logical} whether to show goodness-of-fit value in the result.}

\item{nthread}{\code{integer} Number of cores used to perform computation.
Default 1.}
}
\value{
\link{data.table} contains parameters of projected dose-response curves
for adding one treatment to the other.
}
\description{
Fit projected dose-response curves with \code{E_min} as the viability
of the treatment being added to the other treament at a fixed dose.
}
\examples{
\dontrun{
combo_profiles <- CoreGx::buildComboProfiles(tre, c("HS", "EC50", "E_inf", "viability"))
combo_twowayFit <- fitTwowayZIP(combo_profiles)
}

}
\references{
Yadav, B., Wennerberg, K., Aittokallio, T., & Tang, J. (2015). Searching for Drug Synergy in Complex Dose–Response Landscapes Using an Interaction Potency Model. Computational and Structural Biotechnology Journal, 13, 504–513. https://doi.org/10.1016/j.csbj.2015.09.001
}
