% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fNames.R
\name{fNames<-,PharmacoSet,character,character-method}
\alias{fNames<-,PharmacoSet,character,character-method}
\title{fNames<-}
\usage{
\S4method{fNames}{PharmacoSet,character,character}(object, mDataType) <- value
}
\arguments{
\item{object}{The [`PharmacoSet`] object to update}

\item{mDataType}{[`character`] The molecular data type to update}

\item{value}{A [`character`] vector of the new cell names}
}
\value{
Updated [`PharmacoSet`]
}
\description{
Setter for the feature names of a [`SummarizedExperiment`] in the
  molecularProfiles slot
}
\examples{
data(CCLEsmall)
fNames(CCLEsmall, 'rna') <- fNames(CCLEsmall, 'rna')

}
