% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GDSCsmall}
\alias{GDSCsmall}
\title{Genomics of Drug Sensitivity in Cancer Example PharmacoSet}
\format{PharmacoSet object}
\usage{
data(GDSCsmall)
}
\description{
A small example version of the Genomics of Drug Sensitivity in Cancer Project
PharmacoSet, used in the documentation examples. All credit for the data goes
to the Genomics of Drug Sensitivity in Cancer Project group at the Sanger.This is not a full version of the dataset, most of
of the dataset was removed to make runnable example code. For the full dataset,
please download using the downloadPSet function.
}
\references{
Garnett et al. Systematic identification of genomic markers of drug sensitivity in cancer cells. Nature, 2012.
}
\keyword{datasets}
