% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{molecularProfiles<-}
\alias{molecularProfiles<-}
\title{molecularProfiles<- Generic}
\usage{
molecularProfiles(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to replace molecular profiles in}

\item{mDataType}{The type of molecular data to be updated}

\item{value}{A \code{matrix} with the new profiles}
}
\value{
Updated \code{PharmacoSet}
}
\description{
Generic for molecularProfiles replace method
}
\examples{
data(CCLEsmall)
molecularProfiles(CCLEsmall, "rna") <- molecularProfiles(CCLEsmall, "rna")

}
