% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{molecularProfiles}
\alias{molecularProfiles}
\title{molecularProfiles Generic}
\usage{
molecularProfiles(pSet, mDataType)
}
\arguments{
\item{pSet}{The \code{PharmacoSet} to retrieve molecular profiles from}

\item{mDataType}{the type of molecular data}
}
\value{
a \code{data.frame} with the experiment info
}
\description{
Generic for molecularProfiles method
}
\examples{
data(CCLEsmall)
molecularProfiles(CCLEsmall, "rna")

}
