% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-accessors.R
\name{PharmacoSet-accessors}
\alias{PharmacoSet-accessors}
\alias{drugInfo,PharmacoSet-method}
\alias{drugInfo{`r}
\alias{.local_class`}-method}
\alias{drugInfo}
\alias{drugInfo<-,PharmacoSet,data.frame-method}
\alias{drugInfo<-{`r}
\alias{.local_class`},data.frame-method}
\alias{drugInfo<-}
\alias{cellInfo,PharmacoSet-method}
\alias{cellInfo}
\alias{cellInfo<-,PharmacoSet,data.frame-method}
\alias{cellInfo<-}
\title{Accessing and modifying information in a `PharmacoSet`}
\usage{
\S4method{drugInfo}{PharmacoSet}(object)

\S4method{drugInfo}{PharmacoSet,data.frame}(object) <- value

\S4method{cellInfo}{PharmacoSet}(object)

\S4method{cellInfo}{PharmacoSet,data.frame}(object) <- value
}
\arguments{
\item{object}{A `PharmacoSet` object.}

\item{value}{See details.}

\item{mDataType}{`character(1)` The name of a molecular datatype to access
from the `molecularProfiles` of a `PharmacoSet` object.}

\item{assay}{`character(1)` A valid assay name in the `SummarizedExperiment`
of `@molecularProfiles` of a PharmacoSet object for data type `mDataType`.}

\item{dimension}{See details.}

\item{...}{See details.}
}
\value{
Accessors: See details.

Setters: An updated `PharmacoSet` object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a `PharmacoSet` object.
}
\details{
\subsection{drug slot accessors}{

\strong{drugInfo}: \code{data.frame} Retrieve the drug metadata from a
{PharmacoSet} objects \verb{@drug} slot.
}

\subsection{drug slot accessors}{

\strong{drugInfo}: \code{data.frame} Retrieve the drug metadata from a
{PharmacoSet} objects \verb{@drug} slot.
}

\strong{cellInfo}: \code{data.frame} Metadata for all cell-lines in a \code{PharmacoSet} object.

\strong{cellInfo<-}: assign updated cell-line annotations to the \code{PharmacoSet} object. Arguments:
\itemize{
\item value: a \code{data.frame} object.
}
}
\examples{
drugInf <- drugInfo(CCLEsmall)

drugInfo(`r .local_data`) <- drugInfo(`r .local_data`)

cellInfo(CCLEsmall) <- cellInfo(CCLEsmall)

}
