% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-accessors.R
\name{PharmacoSet-accessors}
\alias{PharmacoSet-accessors}
\alias{drugInfo,PharmacoSet-method}
\alias{drugInfo<-,PharmacoSet,data.frame-method}
\alias{cellInfo,PharmacoSet-method}
\alias{cellInfo}
\title{Accessing and modifying information in a `PharmacoSet`}
\usage{
\S4method{drugInfo}{PharmacoSet}(object)

\S4method{drugInfo}{PharmacoSet,data.frame}(object) <- value

\S4method{cellInfo}{PharmacoSet}(object)
}
\arguments{
\item{object}{A `PharmacoSet` object.}

\item{value}{See details.}

\item{mDataType}{`character(1)` The name of a molecular datatype to access
from the `molecularProfiles` of a `PharmacoSet` object.}

\item{assay}{`character(1)` A valid assay name in the `SummarizedExperiment`
of `@molecularProfiles` of a PharmacoSet object for data type `mDataType`.}

\item{dimension}{See details.}

\item{...}{See details.}
}
\value{
Accessors: See details.

Setters: An updated `PharmacoSet` object, returned invisibly.
}
\description{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a `PharmacoSet` object.
}
\details{
\strong{cellInfo}: \code{data.frame} Metadata for all cell-lines in a \code{PharmacoSet} object.
}
\examples{
data(CCLEsmall)
drugInf <- drugInfo(CCLEsmall)

data(CCLEsmall)
drugInf <- drugInfo(CCLEsmall)

}
