% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-accessors.R
\name{PharmacoSet-accessors}
\alias{PharmacoSet-accessors}
\alias{drugInfo,PharmacoSet-method}
\alias{drugInfo<-,PharmacoSet,data.frame-method}
\title{Accessing and modifying information in a \code{PharmacoSet}}
\usage{
\S4method{drugInfo}{PharmacoSet}(object)

\S4method{drugInfo}{PharmacoSet,data.frame}(object) <- value
}
\arguments{
\item{object}{A \code{PharmacoSet} object.}

\item{value}{See details.}

\item{...}{See details.}
}
\value{
Accessors: See details.

Setters: An updated \code{PharmacoSet} object, returned invisibly.
}
\description{
Accessing and modifying information in a \code{PharmacoSet}
}
\details{
Documentation for the various setters and getters which allow manipulation
of data in the slots of a \code{PharmacoSet} object:
}
\examples{
data(CCLEsmall)
drugInf <- drugInfo(CCLEsmall)

data(CCLEsmall)
drugInf <- drugInfo(CCLEsmall)

}
