% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PharmacoSet.R,
%   R/methods-cellNames.R, R/methods-dateCreated.R, R/methods-drugNames.R,
%   R/methods-fNames.R, R/methods-featureInfo.R, R/methods-mDataNames.R,
%   R/methods-molecularProfiles.R, R/methods-molecularProfilesSlot.R,
%   R/methods-name.R, R/methods-pertNumber.R, R/methods-phenoInfo.R,
%   R/methods-sensNumber.R, R/methods-sensitivityInfo.R,
%   R/methods-sensitivityMeasures.R, R/methods-sensitivityProfiles.R,
%   R/methods-sensitivityRaw.R, R/methods-sensitivitySlot.R
\docType{class}
\name{PharmacoSet-class}
\alias{PharmacoSet-class}
\alias{.PharmacoSet}
\alias{cellNames,PharmacoSet-method}
\alias{cellNames<-,PharmacoSet,character-method}
\alias{dateCreated,PharmacoSet-method}
\alias{sensNumber<-,PharmacoSet,matrix-method}
\alias{drugNames,PharmacoSet-method}
\alias{drugNames<-,PharmacoSet,character-method}
\alias{fNames,PharmacoSet,character-method}
\alias{featureInfo,PharmacoSet,character-method}
\alias{featureInfo<-,PharmacoSet,character,DataFrame-method}
\alias{mDataNames,PharmacoSet-method}
\alias{molecularProfiles,PharmacoSet-method}
\alias{molecularProfiles<-,PharmacoSet,character,character,matrix-method}
\alias{molecularProfiles<-,PharmacoSet,character,missing,matrix-method}
\alias{molecularProfilesSlot,PharmacoSet-method}
\alias{molecularProfilesSlot<-,PharmacoSet,list-method}
\alias{name,PharmacoSet-method}
\alias{name<-,PharmacoSet,character-method}
\alias{pertNumber,PharmacoSet-method}
\alias{pertNumber<-,PharmacoSet,array-method}
\alias{phenoInfo,PharmacoSet-method}
\alias{phenoInfo<-,PharmacoSet,character,DataFrame-method}
\alias{sensNumber,PharmacoSet-method}
\alias{sensitivityInfo,PharmacoSet-method}
\alias{sensitivityInfo<-,PharmacoSet,data.frame-method}
\alias{sensitivityMeasures,PharmacoSet-method}
\alias{sensitivityMeasures<-,PharmacoSet,character-method}
\alias{sensitivityProfiles,PharmacoSet-method}
\alias{sensitivityProfiles<-,PharmacoSet,data.frame-method}
\alias{sensitivityProfiles<-,PharmacoSet,matrix-method}
\alias{sensitivityRaw,PharmacoSet-method}
\alias{sensitivityRaw<-,PharmacoSet,array-method}
\alias{sensitivitySlot,PharmacoSet-method}
\alias{sensitivitySlot<-,PharmacoSet,list-method}
\title{A Class to Contain PharmacoGenomic datasets together with their curations}
\usage{
\S4method{cellNames}{PharmacoSet}(object)

\S4method{cellNames}{PharmacoSet,character}(object) <- value

\S4method{dateCreated}{PharmacoSet}(object)

\S4method{sensNumber}{PharmacoSet,matrix}(object) <- value

\S4method{drugNames}{PharmacoSet}(object)

\S4method{drugNames}{PharmacoSet,character}(object) <- value

\S4method{fNames}{PharmacoSet,character}(object, mDataType)

\S4method{featureInfo}{PharmacoSet,character}(object, mDataType)

\S4method{featureInfo}{PharmacoSet,character,DataFrame}(object, mDataType) <- value

\S4method{mDataNames}{PharmacoSet}(object)

\S4method{molecularProfiles}{PharmacoSet}(object, mDataType, assay)

\S4method{molecularProfiles}{PharmacoSet,character,character,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfiles}{PharmacoSet,character,missing,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfilesSlot}{PharmacoSet}(object)

\S4method{molecularProfilesSlot}{PharmacoSet,list}(object) <- value

\S4method{name}{PharmacoSet}(object)

\S4method{name}{PharmacoSet,character}(object) <- value

\S4method{pertNumber}{PharmacoSet}(object)

\S4method{pertNumber}{PharmacoSet,array}(object) <- value

\S4method{phenoInfo}{PharmacoSet}(object, mDataType)

\S4method{phenoInfo}{PharmacoSet,character,DataFrame}(object, mDataType) <- value

\S4method{sensNumber}{PharmacoSet}(object)

\S4method{sensitivityInfo}{PharmacoSet}(object, dimension, ...)

\S4method{sensitivityInfo}{PharmacoSet,data.frame}(object, dimension, ...) <- value

\S4method{sensitivityMeasures}{PharmacoSet}(object)

\S4method{sensitivityMeasures}{PharmacoSet,character}(object) <- value

\S4method{sensitivityProfiles}{PharmacoSet}(object)

\S4method{sensitivityProfiles}{PharmacoSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{PharmacoSet,matrix}(object) <- value

\S4method{sensitivityRaw}{PharmacoSet}(object)

\S4method{sensitivityRaw}{PharmacoSet,array}(object) <- value

\S4method{sensitivitySlot}{PharmacoSet}(object)

\S4method{sensitivitySlot}{PharmacoSet,list}(object) <- value
}
\arguments{
\item{object}{A \code{PharmacoSet} to extract the raw sensitivity data from}

\item{value}{A \code{list} of new sensitivity slot data for the pSet}

\item{mDataType}{the type of molecular data}

\item{assay}{\code{\link{character}} Name or index of the assay data to return}

\item{dimension}{\code{\link{character}} Optional name of the dimension to extract,
either 'cells' or 'drugs'. Only used if the sensitivity slot contains a
\code{LongTable} object instead of a \code{list}.}

\item{...}{Additional arguments to the rowData or colData
\code{LongTable} methods. Only used if the sensitivity slot contains a
\code{LongTable} object instead of a \code{list}.}
}
\value{
An object of the PharmacoSet class

A vector of the cell names used in the PharmacoSet

Updated \code{\link{PharmacoSet}}

\code{\link{character}} The date the \code{PharmacoSet} was created

The updated \code{PharmacoSet}

A \code{\link{character}} vector containg the names of drugs in the pSet

The updated \code{\link{PharmacoSet}} object

A \code{\link{character}} vector of the feature names

A \code{\link{data.frame}} with the feature annotations for the specified \code{mDataType}

Updated \code{PharmacoSet}

Vector of names of the molecular data types

a \code{\link{matrix}} of data for the given mDataType and assay

Updated \code{\link{PharmacoSet}}

A \code{\link{list}} of \code{SummarizedExperiment} objects, named by molecular data type

\code{\link{character}} The name of the \code{PharmacoSet}

The name of the PharmacoSet

A 3D \code{\link{array}} with the number of perturbation experiments per
drug and cell line, and data type

The updated \code{\link{PharmacoSet}}

a \code{\link{data.frame}} with the phenotype information for the specified molecular data type

The updated \code{PharmacoSet}

A \code{data.frame} with the number of sensitivity experiments per
drug and cell line

a \code{\link{DataFrame}} with the experiment info

Updated \code{PharmacoSet}

A \code{\link{character}} vector of all the available sensitivity measures

A \code{\link{character}} vector of all the available sensitivity measures

a \code{data.frame} with the experiment info

\code{\link{invisible}} Updates the \code{PharmacoSet} object.

\code{\link{invisible}} Updates the \code{PharmacoSet} object.

A \code{array} containing the raw sensitivity data

A copy of the \code{PharmacoSet} containing the updated sensitivty data

A \code{list} of the sensitivity slot contents

A copy of the \code{PharmacoSet} containing the updated sensitivty slot
}
\description{
The PharmacoSet (pSet) class was developed to contain and organise large
PharmacoGenomic datasets, and aid in their metanalysis. It was designed
primarily to allow bioinformaticians and biologists to work with data at the
level of genes, drugs and cell lines, providing a more naturally intuitive
interface and simplifying analyses between several datasets. As such, it was
designed to be flexible enough to hold datasets of two different natures
while providing a common interface. The class can accomidate datasets
containing both drug dose response data, as well as datasets contaning
genetic profiles of cell lines pre and post treatement with compounds, known
respecitively as sensitivity and perturbation datasets.

Get the names of all cell-lines available in a \code{PharmacoSet} object

Update the names of cell lines available in a \code{PharmacoSet} object

Get the data that a \code{PharmacoSet} object was updated

A generic for the sensNumber method

Get the names of all drugs available in a specified \code{PharmacoSet} object

Set the drug names available in a PharmacoSet object

Return the feature names for the specified molecular data type

Get the molecular profile data for the specified molecular data type

Update the molecular profile data for the specified datatype in the specified pSet object

Returns the molecular data names for the \code{PharmacoSet} object

Get the molecular profile data for the specified molecular data type

Update the molecular profile data for the specified datatype in the specified pSet object

Return the name of the PharmacoSet object

Return the name of the \code{PharmacoSet} object

Get the perturbation number for a specified \code{PharmcoSet} object

Set the perturbation number for a specified \code{PharmacoSet} object

Get the phenotype information for a specified molecular datatype

Update the phenotype information for a specified molecular data type in a specified pSet object

Get the sensitivity numbers for a \code{PharmacoSet} object

Get the senstivity information DataFrame from a PharmacoSet object

Set the sensitivityInfo DataFrame in a PharmacoSet object

Get the types of sensitivity measurements from a object object

Get the types of sensitivity measurements available in a PharmacoSet object

Get the sensitivityProfiles data.frame from a PharmacoSet object
}
\section{Methods (by generic)}{
\itemize{
\item \code{cellNames}: Return the cell names used in the dataset

\item \code{cellNames<-}: Update the cell names used in the dataset

\item \code{dateCreated}: Return the date the PharmacoSet was created

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments

\item \code{drugNames}: Return the names of the drugs used in the PharmacoSet

\item \code{drugNames<-}: Update the drug names used in the dataset

\item \code{fNames}: Return the feature names used in the dataset

\item \code{featureInfo}: Return the feature info for the given molecular datatype

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{mDataNames}: Returns the names of molecular data types in a PharmacoSet

\item \code{molecularProfiles}: Return the given type of molecular data from the PharmacoSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the PharmacoSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the PharmacoSet

\item \code{molecularProfilesSlot}: Getter for the molecular profiles slot

\item \code{molecularProfilesSlot<-}: Setter for the molecular profiles slot

\item \code{name}: Return the name of the PharmacoSet

\item \code{name<-}: Return the name of the PharmacoSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{phenoInfo}: Return the experiment info from the given type of
molecular data in PharmacoSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the PharmacoSet

\item \code{sensNumber}: Return the summary of available sensitivity
experiments

\item \code{sensitivityInfo}: Return the drug dose sensitivity experiment info

\item \code{sensitivityInfo<-}: Update the metadata for the treatment response
experiments in the sensitivity slot.

\item \code{sensitivityMeasures}: returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{sensitivityMeasures<-}: returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{sensitivityProfiles}: Return the sensitivity profile summary values for the
treatment response experiment data in the sensitivity slot.

\item \code{sensitivityProfiles<-}: Update the sensitivity profiles for a \code{PharmacoSet}
object.

\item \code{sensitivityProfiles<-}: Update the sensitivity profiles for a \code{PharmacoSet}
object.

\item \code{sensitivityRaw}: Retrive the raw dose and viability data from a pSet

\item \code{sensitivityRaw<-}: Update the raw dose and viability data in a pSet object

\item \code{sensitivitySlot}: Retrieve the contents of the sensitivity slot

\item \code{sensitivitySlot<-}: Set the raw dose and viability data for an pSet and return
and updated copty
}}

\section{Slots}{

\describe{
\item{\code{annotation}}{A \code{list} of annotation data about the PharmacoSet,
including the \code{$name} and the session information for how the object
was creating, detailing the exact versions of R and all the packages used}

\item{\code{molecularProfiles}}{A \code{list} containing \code{SummarizedExperiment}
type object for holding data for RNA, DNA, SNP and CNV
measurements, with associated \code{fData} and \code{pData}
containing the row and column metadata}

\item{\code{cell}}{A \code{data.frame} containing the annotations for all the cell
lines profiled in the data set, across all data types}

\item{\code{drug}}{A \code{data.frame} containg the annotations for all the drugs
profiled in the data set, across all data types}

\item{\code{sensitivity}}{A \code{list} containing all the data for the sensitivity
experiments, including \code{$info}, a \code{data.frame} containing the
experimental info,\code{$raw} a 3D \code{array} containing raw data,
\code{$profiles}, a \code{data.frame} containing sensitivity profiles
statistics, and \code{$n}, a \code{data.frame} detailing the number of
experiments for each cell-drug pair}

\item{\code{perturbation}}{A \code{list} containting \code{$n}, a \code{data.frame}
summarizing the available perturbation data,}

\item{\code{curation}}{A \code{list} containing mappings for \code{$drug},
\code{cell}, \code{tissue} names  used in the data set to universal
identifiers used between different PharmacoSet objects}

\item{\code{datasetType}}{A \code{character} string of 'sensitivity',
'perturbation', or both detailing what type of data can be found in the
PharmacoSet, for proper processing of the data}
}}

\examples{
data(CCLEsmall)
cellNames(CCLEsmall)

data(CCLEsmall)
cellNames(CCLEsmall) <- cellNames(CCLEsmall)

data(CCLEsmall)
dateCreated(CCLEsmall)

data(CCLEsmall)
sensNumber(CCLEsmall) <- sensNumber(CCLEsmall)

data(CCLEsmall)
drugNames(CCLEsmall)

data(CCLEsmall)
drugNames(CCLEsmall) <- drugNames(CCLEsmall)

data(CCLEsmall)
fNames(CCLEsmall, "rna")

data(CCLEsmall)
featInf <- featureInfo(CCLEsmall, "rna")

data(CCLEsmall)
featureInfo(CCLEsmall, "rna") <- featureInfo(CCLEsmall, "rna")

data(CCLEsmall)
mDataNames(CCLEsmall)

data(CCLEsmall)
molProf <- molecularProfiles(CCLEsmall, "rna")

data(CCLEsmall)
molecularProfiles(CCLEsmall, "rna") <- molecularProfiles(CCLEsmall, "rna")

data(CCLEsmall)
molProfSlot <- molecularProfilesSlot(CCLEsmall)

data(CCLEsmall)
molecularProfilesSlot(CCLEsmall) <- molecularProfilesSlot(CCLEsmall)

data(CCLEsmall)
name(CCLEsmall)

data(CCLEsmall)
name(CCLEsmall) <- 'CCLEsmall'

data(CCLEsmall)
pertNumber(CCLEsmall)

data(CCLEsmall)
pertNumber(CCLEsmall) <- pertNumber(CCLEsmall)

data(CCLEsmall)
phenoInf <- phenoInfo(CCLEsmall, mDataType="rna")

data(CCLEsmall)
phenoInfo(CCLEsmall, mDataType='rna') <- phenoInfo(CCLEsmall, mDataType='rna')

data(CCLEsmall)
sensNumber(CCLEsmall)

data(CCLEsmall)
sensInf <- sensitivityInfo(CCLEsmall)

data(CCLEsmall)
sensitivityInfo(CCLEsmall) <- sensitivityInfo(CCLEsmall)

data(CCLEsmall)
sensMeas <- sensitivityMeasures(CCLEsmall)

data(CCLEsmall)
sensMeas <- sensitivityMeasures(CCLEsmall)

data(CCLEsmall)
sensProf <- sensitivityProfiles(CCLEsmall)

data(GDSCsmall)
sensitivityProfiles(GDSCsmall) <- sensitivityProfiles(GDSCsmall)

data(GDSCsmall)
sensitivityProfiles(GDSCsmall) <- sensitivityProfiles(GDSCsmall)

data(CCLEsmall)
sensitivityRaw(CCLEsmall)

data(CCLEsmall)
sensitivityRaw(CCLEsmall) <- sensitivityRaw(CCLEsmall)

data(CCLEsmall)
sensitivitySlot(CCLEsmall)

data(CCLEsmall)
sensitivitySlot(CCLEsmall) <- sensitivitySlot(CCLEsmall)

}
