% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{class}
\name{PharmacoSet-class}
\alias{PharmacoSet-class}
\alias{.PharmacoSet}
\alias{cellInfo,PharmacoSet-method}
\alias{drugInfo,PharmacoSet-method}
\alias{drugInfo<-,PharmacoSet,data.frame-method}
\alias{phenoInfo,PharmacoSet-method}
\alias{phenoInfo<-,PharmacoSet,character,DataFrame-method}
\alias{molecularProfiles,PharmacoSet-method}
\alias{molecularProfiles<-,PharmacoSet,character,character,matrix-method}
\alias{molecularProfiles<-,PharmacoSet,character,missing,matrix-method}
\alias{featureInfo,PharmacoSet-method}
\alias{featureInfo<-,PharmacoSet,character,DataFrame-method}
\alias{sensitivityInfo,PharmacoSet-method}
\alias{sensitivityInfo<-,PharmacoSet,data.frame-method}
\alias{sensitivityProfiles,PharmacoSet-method}
\alias{sensitivityProfiles<-,PharmacoSet,data.frame-method}
\alias{sensitivityProfiles<-,PharmacoSet,matrix-method}
\alias{sensitivityMeasures,PharmacoSet-method}
\alias{drugNames,PharmacoSet-method}
\alias{drugNames<-,PharmacoSet,character-method}
\alias{cellNames,PharmacoSet-method}
\alias{cellNames<-,PharmacoSet,character-method}
\alias{fNames,PharmacoSet,character-method}
\alias{dateCreated,PharmacoSet-method}
\alias{name,PharmacoSet-method}
\alias{pertNumber,PharmacoSet-method}
\alias{sensNumber,PharmacoSet-method}
\alias{pertNumber<-,PharmacoSet,array-method}
\alias{sensNumber<-,PharmacoSet,matrix-method}
\title{A Class to Contain PharmacoGenomic datasets together with their curations}
\usage{
\S4method{cellInfo}{PharmacoSet}(object)

\S4method{drugInfo}{PharmacoSet}(object)

\S4method{drugInfo}{PharmacoSet,data.frame}(object) <- value

\S4method{phenoInfo}{PharmacoSet}(object, mDataType)

\S4method{phenoInfo}{PharmacoSet,character,DataFrame}(object, mDataType) <- value

\S4method{molecularProfiles}{PharmacoSet}(object, mDataType, assay)

\S4method{molecularProfiles}{PharmacoSet,character,character,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfiles}{PharmacoSet,character,missing,matrix}(object, mDataType, assay) <- value

\S4method{featureInfo}{PharmacoSet}(object, mDataType)

\S4method{featureInfo}{PharmacoSet,character,DataFrame}(object, mDataType) <- value

\S4method{sensitivityInfo}{PharmacoSet}(object)

\S4method{sensitivityInfo}{PharmacoSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{PharmacoSet}(object)

\S4method{sensitivityProfiles}{PharmacoSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{PharmacoSet,matrix}(object) <- value

\S4method{sensitivityMeasures}{PharmacoSet}(object)

\S4method{drugNames}{PharmacoSet}(object)

\S4method{drugNames}{PharmacoSet,character}(object) <- value

\S4method{cellNames}{PharmacoSet}(object)

\S4method{cellNames}{PharmacoSet,character}(object) <- value

\S4method{fNames}{PharmacoSet,character}(object, mDataType)

\S4method{dateCreated}{PharmacoSet}(object)

\S4method{name}{PharmacoSet}(object)

\S4method{pertNumber}{PharmacoSet}(object)

\S4method{sensNumber}{PharmacoSet}(object)

\S4method{pertNumber}{PharmacoSet,array}(object) <- value

\S4method{sensNumber}{PharmacoSet,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{PharmacoSet}}

\item{value}{A new \code{data.frame} with the number of sensitivity experiments per drug and cell line}

\item{mDataType}{The molecular data type to return feature names for}

\item{assay}{\code{character} Name or index of the assay data to return}

\item{...}{\code{list} Fall through parameters to other functions. Needed to add addtional arguments to a generic.}
}
\value{
An object of the PharmacoSet class

a \code{data.frame} with the cell annotations

a \code{data.frame} with the experiment info

The updated \code{PharmacoSet}

a \code{matrix} of data for the given mDataType and assay

Updated \code{PharmacoSet}

a \code{data.frame} with the experiment info

Updated \code{PharmacoSet}

a \code{DataFrame} with the experiment info

Updated \code{PharmacoSet}

a \code{data.frame} with the experiment info

Updated \code{PharmacoSet}

A \code{character} vector of all the available sensitivity measures

A vector of the cell names used in the PharmacoSet

Updated \code{PharmacoSet}

A \code{character} vector of the feature names

The date the PharmacoSet was created

The name of the PharmacoSet

A 3D \code{array} with the number of perturbation experiments per drug and cell line, and data type

A \code{data.frame} with the number of sensitivity experiments per drug and cell line

The updated \code{PharmacoSet}

The updated \code{PharmacoSet}
}
\description{
The PharmacoSet (pSet) class was developed to contain and organise large 
PharmacoGenomic datasets, and aid in their metanalysis. It was designed 
primarily to allow bioinformaticians and biologists to work with data at the 
level of genes, drugs and cell lines, providing a more naturally intuitive 
interface and simplifying analyses between several datasets. As such, it was 
designed to be flexible enough to hold datasets of two different natures 
while providing a common interface. The class can accomidate datasets 
containing both drug dose response data, as well as datasets contaning 
genetic profiles of cell lines pre and post treatement with compounds, known 
respecitively as sensitivity and perturbation datasets.

Return cell line metadata from a object

Generic for phenoInfo method

Generic for phenoInfo replace method

Generic for molecularProfiles method

Generic for molecularProfiles replace method

Generic for featureInfo method

Generic for featureInfo replace method

Get the senstivity information DataFrame from a object object

Set the sensitivityInfo DataFrame in a object object

Get the sensitivityProfiles data.frame from a object object

Set the sensitivityProfiles in a object object

Get the types of sensitivity measurements from a object object

A generic for the cellNames method

A generic for the cellNames replacement method

Return the feature names for the specified molecular data type

A generic for the dateCreated method

Retrun the name of the PharmacoSet object

A generic for the pertNumber method

A generic for the sensNumber method

A generic for the pertNumber method

A generic for the sensNumber method
}
\section{Methods (by generic)}{
\itemize{
\item \code{cellInfo}: 

\item \code{drugInfo}: Returns the annotations for all the drugs tested in the PharmacoSet

\item \code{drugInfo<-}: Update the drug annotations

\item \code{phenoInfo}: Return the experiment info from the given type of molecular data in PharmacoSet

\item \code{phenoInfo<-}: Update the given type of molecular data experiment info in the PharmacoSet

\item \code{molecularProfiles}: Return the given type of molecular data from the PharmacoSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the PharmacoSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the PharmacoSet

\item \code{featureInfo}: Return the feature info for the given molecular data

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{sensitivityInfo}: Return the drug dose sensitivity experiment info

\item \code{sensitivityInfo<-}: Update the sensitivity experiment info

\item \code{sensitivityProfiles}: Return the phenotypic data for the drug dose sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityMeasures}: returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{drugNames}: Return the names of the drugs used in the PharmacoSet

\item \code{drugNames<-}: Update the drug names used in the dataset

\item \code{cellNames}: Return the cell names used in the dataset

\item \code{cellNames<-}: Update the cell names used in the dataset

\item \code{fNames}: Return the feature names used in the dataset

\item \code{dateCreated}: Return the date the PharmacoSet was created

\item \code{name}: Return the name of the PharmacoSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{sensNumber}: Return the summary of available sensitivity
experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments
}}

\section{Slots}{

\describe{
\item{\code{annotation}}{A \code{list} of annotation data about the PharmacoSet,
including the \code{$name} and the session information for how the object
was creating, detailing the exact versions of R and all the packages used}

\item{\code{molecularProfiles}}{A \code{list} containing \code{SummarizedExperiment} 
type object for holding data for RNA, DNA, SNP and CNV 
measurements, with associated \code{fData} and \code{pData} 
containing the row and column metadata}

\item{\code{cell}}{A \code{data.frame} containing the annotations for all the cell 
lines profiled in the data set, across all data types}

\item{\code{drug}}{A \code{data.frame} containg the annotations for all the drugs 
profiled in the data set, across all data types}

\item{\code{sensitivity}}{A \code{list} containing all the data for the sensitivity 
experiments, including \code{$info}, a \code{data.frame} containing the 
experimental info,\code{$raw} a 3D \code{array} containing raw data, 
\code{$profiles}, a \code{data.frame} containing sensitivity profiles 
statistics, and \code{$n}, a \code{data.frame} detailing the number of 
experiments for each cell-drug pair}

\item{\code{perturbation}}{A \code{list} containting \code{$n}, a \code{data.frame} 
summarizing the available perturbation data,}

\item{\code{curation}}{A \code{list} containing mappings for \code{$drug}, 
\code{cell}, \code{tissue} names  used in the data set to universal 
identifiers used between different PharmacoSet objects}

\item{\code{datasetType}}{A \code{character} string of 'sensitivity', 
'perturbation', or both detailing what type of data can be found in the 
PharmacoSet, for proper processing of the data}
}}

\examples{
data(CCLEsmall)
cellInfo(CCLEsmall)

data(CCLEsmall)
phenoInfo(CCLEsmall, mDataType="rna")

data(CCLEsmall)
phenoInfo(CCLEsmall, mDataType="rna") <- phenoInfo(CCLEsmall, mDataType="rna")

data(CCLEsmall)
molecularProfiles(CCLEsmall, "rna")

data(CCLEsmall)
molecularProfiles(CCLEsmall, "rna") <- molecularProfiles(CCLEsmall, "rna")

data(CCLEsmall)
featureInfo(CCLEsmall, "rna")

data(CCLEsmall)
featureInfo(CCLEsmall, "rna") <- featureInfo(CCLEsmall, "rna")

data(CCLEsmall)
sensitivityInfo(CCLEsmall)

data(CCLEsmall)
sensitivityInfo(CCLEsmall) <- sensitivityInfo(CCLEsmall)

data(CCLEsmall)
sensitivityProfiles(CCLEsmall)

data(CCLEsmall)
sensitivityProfiles(CCLEsmall) <- sensitivityProfiles(CCLEsmall)

data(CCLEsmall)
sensitivityMeasures(CCLEsmall)

data(CCLEsmall)
cellNames(CCLEsmall)

data(CCLEsmall)
cellNames(CCLEsmall) <- cellNames(CCLEsmall)

data(CCLEsmall)
fNames(CCLEsmall, "rna")

data(CCLEsmall)
dateCreated(CCLEsmall)

data(CCLEsmall)
name(CCLEsmall)

data(CCLEsmall)
pertNumber(CCLEsmall)

data(CCLEsmall)
sensNumber(CCLEsmall)

data(CCLEsmall)
pertNumber(CCLEsmall) <- pertNumber(CCLEsmall)

data(CCLEsmall)
sensNumber(CCLEsmall) <- sensNumber(CCLEsmall)

}
