% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-class.R
\name{PharmacoSet2}
\alias{PharmacoSet2}
\title{New implementation of the CoreSet constructor to support MAE and TRE. This
constructor will be swapped with the original \code{CoreSet} constructor as
part of an overhaul of the CoreSet class structure.}
\usage{
PharmacoSet2(
  name = "emptySet",
  treatment = data.frame(),
  sample = data.frame(),
  molecularProfiles = MultiAssayExperiment(),
  treatmentResponse = LongTable(),
  curation = list(sample = data.frame(), treatment = data.frame())
)
}
\description{
\preformatted{@param name A `character(1)` vector with the `\{class_\}` objects name.
@param treatment A `data.frame` with treatment level metadata. \{tx_\}
@param sample A `data.frame` with sample level metadata for the union
    of samples in `treatmentResponse` and `molecularProfiles`. \{sx_\}
@param molecularProfiles A `MultiAssayExperiment` containing one 
    `SummarizedExperiment` object for each molecular data type.
@param treatmentResponse A `LongTable` or `LongTableDataMapper` object
    containing all treatment response data associated with the `\{class_\}`
    object.
@param curation A `list(2)` object with two items named `treatment` and
    `sample` with mappings from publication identifiers to standardized
    identifiers for both annotations, respectively.
@param datasetType A `character(1)` One of 
}
}
\details{
\preformatted{@return A `CoreSet` object storing standardized and curated treatment
    response and multiomic profile data associated with a given publication.

@md
@importFrom MultiAssayExperiment MultiAssayExperiment
}
}
