% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{phenoInfo<-}
\alias{phenoInfo<-}
\title{phenoInfo<- Generic}
\usage{
phenoInfo(object, mDataType) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to retrieve molecular experiment annotations from}

\item{mDataType}{the type of molecular data}

\item{value}{a \code{data.frame} with the new experiment annotations}
}
\value{
The updated \code{PharmacoSet}
}
\description{
Generic for phenoInfo replace method
}
\examples{

data(CCLEsmall)
phenoInfo(CCLEsmall, mDataType="rna") <- phenoInfo(CCLEsmall, mDataType="rna")

}
