% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityInfo<-}
\alias{sensitivityInfo<-}
\title{sensitivityInfo<- Generic}
\usage{
sensitivityInfo(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to update}

\item{value}{A \code{data.frame} with the new sensitivity annotations}
}
\value{
Updated \code{PharmacoSet}
}
\description{
A generic for the sensitivityInfo replacement method
}
\examples{
data(CCLEsmall)
sensitivityInfo(CCLEsmall) <- sensitivityInfo(CCLEsmall)

}
