% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityMeasures}
\alias{sensitivityMeasures}
\title{sensitivityMeasures Generic}
\usage{
sensitivityMeasures(pSet)
}
\arguments{
\item{pSet}{The \code{PharmacoSet}}
}
\value{
A \code{character} vector of all the available sensitivity measures
}
\description{
A generic for the sensitivityMeasures  method
}
\examples{
data(CCLEsmall)
sensitivityMeasures(CCLEsmall)

}
