% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\name{sensitivityProfiles<-}
\alias{sensitivityProfiles<-}
\title{sensitivityProfiles<- Generic}
\usage{
sensitivityProfiles(object) <- value
}
\arguments{
\item{object}{The \code{PharmacoSet} to update}

\item{value}{A \code{data.frame} with the new sensitivity profiles. If a matrix object is passed in, converted to data.frame before assignment}
}
\value{
Updated \code{PharmacoSet}
}
\description{
A generic for the sensitivityProfiles replacement method
}
\examples{
data(CCLEsmall)
sensitivityProfiles(CCLEsmall) <- sensitivityProfiles(CCLEsmall)

}
