% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSet-class.R
\name{show,PharmacoSet-method}
\alias{show,PharmacoSet-method}
\title{Show a PharamcoSet}
\usage{
\S4method{show}{PharmacoSet}(object)
}
\arguments{
\item{object}{\code{PharmacoSet}}
}
\value{
Prints the PharmacoSet object to the output stream, and returns
invisible NULL.

@importFrom CoreGx show
@importFrom methods callNextMethod
}
\description{
Show a PharamcoSet
}
\examples{
data(CCLEsmall)
CCLEsmall

}
