% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{methods}
\name{show,PharmacoSet-method}
\alias{show,PharmacoSet-method}
\title{Show a PharamcoSet}
\usage{
\S4method{show}{PharmacoSet}(object)
}
\arguments{
\item{object}{\code{PharmacoSet}}
}
\value{
Prints the PharmacoSet object to the output stream, and returns invisible NULL.
}
\description{
Show a PharamcoSet
}
\examples{
data(CCLEsmall)
CCLEsmall

}
