% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\docType{methods}
\name{show,PharmacoSig-method}
\alias{show,PharmacoSig-method}
\title{Show PharmacoGx Signatures}
\usage{
\S4method{show}{PharmacoSig}(object)
}
\arguments{
\item{object}{\code{PharmacoSig}}
}
\value{
Prints the PharmacoGx Signatures object to the output stream, and returns invisible NULL.
}
\description{
Show PharmacoGx Signatures
}
\examples{
data(GDSCsmall)
drug.sensitivity <- drugSensitivitySig(GDSCsmall, mDataType="rna", 
             nthread=1, features = fNames(GDSCsmall, "rna")[1])
drug.sensitivity

}
