% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{methods}
\name{[,PharmacoSet-method}
\alias{[,PharmacoSet-method}
\title{`[`}
\usage{
\S4method{[}{PharmacoSet}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{numeric}

\item{i}{numeric}

\item{j}{numeric}

\item{...}{further arguments}

\item{drop}{A boolean flag of whether to drop single dimensions or not}
}
\description{
`[`
}

