% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoSetClass.R
\docType{methods}
\name{[,PharmacoSet-method}
\alias{[,PharmacoSet-method}
\title{`[`}
\usage{
\S4method{[}{PharmacoSet}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{PSet}

\item{i}{Cell lines to keep in PSet}

\item{j}{Drugs to keep in PSet}

\item{...}{further arguments}

\item{drop}{A boolean flag of whether to drop single dimensions or not}
}
\value{
Returns the subsetted PSet
}
\description{
`[`
}
