% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-subset.R
\name{subset,LongTable-method}
\alias{subset,LongTable-method}
\title{Subset method for a LongTable object.}
\usage{
\S4method{subset}{LongTable}(x, i, j, assays, reindex = TRUE)
}
\arguments{
\item{x}{[`LongTable`] The object to subset.}

\item{i}{[`character`], [`numeric`], [`logical`] or [`expression`]
 Character: pass in a character vector of drug names, which will subset the
     object on all row id columns matching the vector.

 Numeric or Logical: these select based on the rowKey from the `rowData`
     method for the `LongTable`.

 Expression: Accepts valid query statements to the `data.table` i parameter,
     this can be used to make complex queries using the `data.table` API
     for the `rowData` data.table.}

\item{j}{[`character`], [`numeric`], [`logical`] or [`expression`]
 Character: pass in a character vector of drug names, which will subset the
     object on all drug id columns matching the vector.

 Numeric or Logical: these select based on the rowID from the `rowData`
     method for the `LongTable`.

 Expression: Accepts valid query statements to the `data.table` i parameter,
     this can be used to make complex queries using the `data.table` API
     for the `rowData` data.table.}

\item{reindex}{[`logical`] Should the col/rowKeys be remapped after subsetting.
defaults to TRUE, but can slow down subsets for large LongTables.}

\item{values}{[`character`, `numeric` or `logical`] Optional list of value
names to subset. Can be used to subset the dataList column further,
returning only the selected items in the new LongTable.}
}
\value{
[`LongTable`] A new `LongTable` object subset based on the specified
     parameters.
}
\description{
Allows use of the colData and rowData `data.table` objects to query based on
 rowID and colID, which is then used to subset all value data.tables stored
 in the dataList slot.
}
\details{
This function is endomorphic, it always returns a LongTable object.
}
